% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_anova_assumption.R
\name{add_anova_assumption}
\alias{add_anova_assumption}
\title{Add ANOVA Estimated Marginal Means Interaction Sheet}
\usage{
add_anova_assumption(
  wb,
  sheet_name,
  anova_object,
  digits = 3,
  name = NULL,
  within = FALSE
)
}
\arguments{
\item{wb}{An \code{openxlsx} workbook object to add the sheet to.}

\item{sheet_name}{A character string reflecting the name of the worksheet.}

\item{anova_object}{The ANOVA object from \code{anova_wrapper}.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{name}{An optional character string reflecting the name of the study.}
}
\description{
Add ANOVA Estimated Marginal Means Interaction Sheet
}
