% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_anova_emmc.R
\name{add_anova_emmc}
\alias{add_anova_emmc}
\title{Add ANOVA Estimated Marginal Means Contrasts Sheet}
\usage{
add_anova_emmc(wb, sheet_name, anova_object, digits = 3, name = NULL)
}
\arguments{
\item{wb}{An \code{openxlsx} workbook object to add the sheet to.}

\item{sheet_name}{A character string reflecting the name of the worksheet.}

\item{anova_object}{The ANOVA object from \code{anova_wrapper}.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{name}{An optional character string reflecting the name of the study.}
}
\description{
Add ANOVA Estimated Marginal Means Contrasts Sheet
}
