% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_anova_report.R
\name{add_anova_report}
\alias{add_anova_report}
\title{Add ANOVA Reportable Table Format}
\usage{
add_anova_report(
  wb,
  sheet_name,
  anova_object,
  name = NULL,
  digits = 3,
  report_es
)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{anova_object}{}

\item{name}{}

\item{digits}{}
}
\description{
Add ANOVA Reportable Table Format
}
