% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_anova_summary.R
\name{add_anova_sum}
\alias{add_anova_sum}
\title{Add ANOVA Summary Sheet}
\usage{
add_anova_sum(
  wb,
  sheet_name,
  anova_object,
  name = NULL,
  digits = 3,
  summary_type = 1
)
}
\arguments{
\item{wb}{An \code{openxlsx} workbook object to add the sheet to.}

\item{sheet_name}{A character string reflecting the name of the worksheet.}

\item{anova_object}{The ANOVA object from \code{anova_wrapper}.}

\item{name}{An optional character string reflecting the name of the study.}

\item{digits}{A numeric value indicating the number of digits to round to.}
}
\description{
Add ANOVA Summary Sheet
}
