% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plssem_diagnostic.R
\name{add_plssem_diagnostic}
\alias{add_plssem_diagnostic}
\title{Affix PLS-SEM Model Diagnostic Workbook Sheet}
\usage{
add_plssem_diagnostic(wb, plssem_mod, sheet_name, name = NULL, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Affix PLS-SEM Model Diagnostic Workbook Sheet
}
