% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plssem_loadings_weight.R
\name{add_plssem_load_wgt}
\alias{add_plssem_load_wgt}
\title{Add PLS-SEM Model Loadings & Weight Workbook Sheet}
\usage{
add_plssem_load_wgt(wb, plssem_mod, sheet_name, digits = 3)
}
\arguments{
\item{wb}{}

\item{plssem_mod}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
a
}
