% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plssem_long_table.R
\name{add_plssem_long_table}
\alias{add_plssem_long_table}
\title{Title}
\usage{
add_plssem_long_table(wb, plssem_mod, sheet_name, digits = 3)
}
\arguments{
\item{wb}{}

\item{plssem_mod}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Title
}
