% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plssem_table_wide.R
\name{add_plssem_wide_table}
\alias{add_plssem_wide_table}
\title{Title}
\usage{
add_plssem_wide_table(wb, plssem_mod, sheet_name, name = NULL, digits = 3)
}
\arguments{
\item{wb}{}

\item{plssem_mod}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Title
}
