% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ttest_assumptions.R
\name{add_ttest_assumptions}
\alias{add_ttest_assumptions}
\title{Apply T-Test Assumption Format}
\usage{
add_ttest_assumptions(wb, sheet_name, ttest_object, name = NULL, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{ttest_object}{}

\item{name}{}

\item{digits}{}
}
\description{
Apply T-Test Assumption Format
}
