% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ttest_report.R
\name{add_ttest_report}
\alias{add_ttest_report}
\title{Add T-Test Reportable Table Sheet}
\usage{
add_ttest_report(
  wb,
  sheet_name,
  ttest_object,
  name = NULL,
  digits = 3,
  report_variance = "sd",
  report_test = "welch"
)
}
\arguments{
\item{wb}{An \code{openxlsx} workbook object to apply the format to.}

\item{sheet_name}{A character string reflecting the sheet.}

\item{ttest_object}{The t-test object from \code{wrap_ttest}.}

\item{name}{An optional character string indicating the name of the study.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{report_variance}{A string value or vector of strings indicating which
variance measure to report. Options include \code{"sd"} for standard
deviation, \code{"se"} for standard error, or \code{c("sd", "se")} for
both. The default is standard deviation \code{"sd"}.}

\item{report_test}{A string value indicating which t-test statistic to
report. Options include \code{"student"} for Student's t-test,
\code{"welch"} for Welch's t-test, and \code{"wilcox"} for Wilcoxon
Rank-Sum W test / Mann-Whitney U test. The default is \code{"welch"}.}
}
\description{
Add T-Test Reportable Table Sheet
}
