% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ttest_summary.R
\name{add_ttest_summary}
\alias{add_ttest_summary}
\title{Add T-Test Summary Sheet}
\usage{
add_ttest_summary(wb, sheet_name, ttest_object, name = NULL, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{ttest_object}{}

\item{name}{}

\item{digits}{}
}
\description{
Add T-Test Summary Sheet
}
