% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affix_plssem_diagnostic.R
\name{affix_plssem_diagnostic}
\alias{affix_plssem_diagnostic}
\title{Affix PLS-SEM Model Diagnostic Workbook Sheet}
\usage{
affix_plssem_diagnostic(wb, plssem_mod, sheet_name, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Affix PLS-SEM Model Diagnostic Workbook Sheet
}
