% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affix_plssem_loadings_weight.R
\name{affix_plssem_load_wgt}
\alias{affix_plssem_load_wgt}
\title{Affix PLS-SEM Model Loadings & Weight Workbook Sheet}
\usage{
affix_plssem_load_wgt(wb, plssem_mod, sheet_name, digits = 3)
}
\arguments{
\item{wb}{}

\item{plssem_mod}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Affix PLS-SEM Model Loadings & Weight Workbook Sheet
}
