% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affix_plssem_pathways.R
\name{affix_plssem_pathways}
\alias{affix_plssem_pathways}
\title{Affix PLS-SEM Model Pathways Workbook Sheet}
\usage{
affix_plssem_pathways(wb, plssem_mod, sheet_name, digits = 3)
}
\arguments{
\item{wb}{}

\item{plssem_mod}{}

\item{sheet_name}{}

\item{digits}{}
}
\description{
Affix PLS-SEM Model Pathways Workbook Sheet
}
