% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_anova_emmc_format.R
\name{apply_anova_emmc_format}
\alias{apply_anova_emmc_format}
\title{Apply ANOVA EMM Contrasts Format}
\usage{
apply_anova_emmc_format(
  wb,
  sheet,
  mod_row,
  mod_col,
  start_col,
  start_row,
  digits = 3
)
}
\arguments{
\item{wb}{An \code{openxlsx} workbook object to apply the format to.}

\item{sheet}{A character string reflecting the sheet to apply the format to.}

\item{mod_row}{A numeric value reflecting the n row of the table.}

\item{mod_col}{A numeric value reflecting the n col of the table.}

\item{start_col}{A numeric value reflecting the starting column of the table.}

\item{start_row}{A numeric value reflecting the starting row of the table.}

\item{digits}{A numeric value indicating the number of digits to round to.}
}
\description{
Apply ANOVA EMM Contrasts Format
}
