% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_anova_summary_format.R
\name{apply_anova_summary_formatter}
\alias{apply_anova_summary_formatter}
\title{Apply ANOVA Summary Formatting}
\usage{
apply_anova_summary_formatter(
  wb,
  sheet,
  mod_row,
  mod_col,
  start_col,
  start_row,
  digits = 3,
  correction = FALSE,
  bs_row = NULL,
  ws_row = NULL
)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{mod_row}{}

\item{mod_col}{}

\item{start_col}{}

\item{start_row}{}

\item{digits}{}
}
\description{
Apply ANOVA Summary Formatting
}
