% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_performance_format.R
\name{apply_perf_formatter}
\alias{apply_perf_formatter}
\title{Apply Performance Formatting}
\usage{
apply_perf_formatter(wb, sheet, df, start_col, start_row, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{df}{}

\item{start_col}{}

\item{start_row}{}
}
\description{
Apply Performance Formatting
}
