% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_plssem_long_table_formatter.R
\name{apply_plssem_long_table_formatter}
\alias{apply_plssem_long_table_formatter}
\title{Apply PLS-SEM Effects Table Formatting}
\usage{
apply_plssem_long_table_formatter(
  wb,
  sheet,
  df,
  start_col,
  start_row,
  digits = 3
)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{df}{}

\item{start_col}{}

\item{start_row}{}

\item{digits}{}
}
\description{
Apply PLS-SEM Effects Table Formatting
}
