% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_plssem_path_formatter.R
\name{apply_plssem_path_formatter}
\alias{apply_plssem_path_formatter}
\title{Apply Performance Formatting}
\usage{
apply_plssem_path_formatter(
  wb,
  sheet,
  df,
  ci_col = 6:7,
  start_col,
  start_row,
  digits = 3,
  direct_nrow = NULL,
  indirect_nrow = NULL
)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{df}{}

\item{start_col}{}

\item{start_row}{}
}
\description{
Apply Performance Formatting
}
