% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_plssem_text_format.R
\name{apply_text_format}
\alias{apply_text_format}
\title{Apply Text Table Formatting}
\usage{
apply_text_format(wb, sheet, df, start_col, start_row, digits = 3)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{df}{}

\item{start_col}{}

\item{start_row}{}
}
\description{
Apply Text Table Formatting
}
