% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_ttest_desc_text_format.R
\name{apply_ttest_desc_text_format}
\alias{apply_ttest_desc_text_format}
\title{Apply T-Test Descriptive Text Formatting}
\usage{
apply_ttest_desc_text_format(
  wb,
  sheet,
  start_col,
  start_row,
  mod_row,
  mod_col,
  digits = 3
)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{start_col}{}

\item{start_row}{}

\item{df}{}
}
\description{
Apply T-Test Descriptive Text Formatting
}
