% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_plssem_wide_table_formatter.R
\name{apply_wide_table_formatter}
\alias{apply_wide_table_formatter}
\title{Title}
\usage{
apply_wide_table_formatter(
  wb,
  sheet,
  df,
  paths,
  start_col,
  start_row,
  digits = 3
)
}
\arguments{
\item{wb}{}

\item{sheet}{}

\item{df}{}

\item{paths}{}

\item{start_col}{}

\item{start_row}{}

\item{digits}{}
}
\description{
Title
}
