% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_fit.R
\name{aug_fit}
\alias{aug_fit}
\title{Augmented Data Accuracy Evaluation}
\usage{
aug_fit(aug_vals, org_vals)
}
\arguments{
\item{aug_vals}{}

\item{org_vals}{}
}
\description{
Calculate the fit metrics of the augmented data against the observed data.
}
