% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_master.R
\name{build_master}
\alias{build_master}
\title{Build a Master Script From Folder}
\usage{
build_master(dir, master_name = "Master", author = "", layout = "compact")
}
\arguments{
\item{dir}{The folder structure which contains the scripts to build upon.}

\item{master_name}{The file name which should be written.}

\item{author}{Authors name to be put in the header.}

\item{layout}{The layout which should be written to the master file. Available
are:
\itemize{
\item "compact": The header and rebuilt master section followed by one script block
which runs all scripts. (default)
\item "full":    The whole folder structure with an overview, a section which runs
all scripts and single blocks for every folder and script.
}}
}
\value{
Returns the script as character vector and saves it as master file.
}
\description{
\code{\link[=build_master]{build_master()}} reads a given folder structure, which contains scripts, and builds
a master file. The file which gets written depends on the selected layout.
}
\details{
The function works with folder structures that look like this:

root/

\if{html}{\out{<div class="sourceCode">}}\preformatted{ subfolder1/

     script1.R

     script2.R

     ....R

 subfolder2/

     script3.R

     script4.R

     ....R

 .../

     ....R
}\if{html}{\out{</div>}}
}
\examples{
# Example export file paths
# NOTE: These tempfiles are only for the examples. In reality you just call the
# main function and put in your desired path and name directly.
temp_file <- tempfile(fileext = ".Rmd")

# Example master
build_master(dir         = dirname(temp_file),
             master_name = basename(temp_file))

# Example full master
build_master(dir         = dirname(temp_file),
             master_name = basename(temp_file),
             layout      = "full")

# Manual cleanup for example
unlink(temp_file)

}
\seealso{
RStudio helper functions: \code{\link[=run_script]{run_script()}}, \code{\link[=run_folder]{run_folder()}}, \code{\link[=run_project]{run_project()}},
\code{\link[=run_project_parallel]{run_project_parallel()}}
}
