% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anova.R
\name{calc_anova}
\alias{calc_anova}
\title{Calculate Between-Subjects ANOVA}
\usage{
calc_anova(
  data,
  outcome,
  between_vars,
  covariates = NULL,
  interactions = FALSE,
  posthoc_correction = "tukey",
  es_type = "hedges",
  conf_level = 0.95,
  emm_weights = TRUE
)
}
\arguments{
\item{data}{}

\item{outcome}{}

\item{between_vars}{}

\item{covariates}{}

\item{interactions}{}

\item{posthoc_correction}{}

\item{es_type}{}

\item{conf_level}{}

\item{emm_weights}{}

\item{g_correction}{}
}
\description{
Calculate Between-Subjects ANOVA
}
