% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anova_assumptions.R
\name{calc_aov_assumptions}
\alias{calc_aov_assumptions}
\title{Calculate ANOVA Assumptions Tests}
\usage{
calc_aov_assumptions(data, outcome, groups, aov_result)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcome}{Character indicating the outcome/dependent variable.}

\item{groups}{Character indicating the factor grouping variable.}

\item{aov_result}{The ANOVA model specified via \code{stats::aov()}.}
}
\description{
Calculate ANOVA Assumptions Tests
}
