% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anova_emm.R
\name{calc_aov_emm}
\alias{calc_aov_emm}
\title{Calculate ANOVA Estimated Marginal Means}
\usage{
calc_aov_emm(
  data,
  groups,
  outcome,
  between_vars = NULL,
  within_vars = NULL,
  covariates = NULL,
  aov_result,
  aov_summary,
  conf_level,
  ph_method,
  es_type,
  g_correction,
  paired = FALSE
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{groups}{Character indicating the factor grouping variable.}

\item{outcome}{Character indicating the outcome/dependent variable.}

\item{between_vars}{An optional character indicating the between groups
factor grouping variable.}

\item{within_vars}{Characters indicating the repeated measures variables.}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{aov_result}{The ANOVA model specified via \code{stats::aov()}.}

\item{aov_summary}{ANOVA summary dataframe created from
\code{calc_aov_summary()}.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{ph_method}{The post-hoc pairwise comparison correction method. Options
include \code{"tukey"}, \code{"scheffe"} , \code{"sidak"},
\code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and \code{"none"}.
The default is \code{"tukey"}. See below for additional information.}

\item{paired}{A logical reflecting whether the pairwise comparison is between
(FALSE) or within (TRUE) subjects. If within, then the pair is tested
whether it is truly within by iterating through the unique between subjects
combinations. \code{bs_unique} must be specified.}
}
\description{
Calculate ANOVA Estimated Marginal Means
}
