% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anova_summary.R
\name{calc_aov_summary}
\alias{calc_aov_summary}
\title{Calculate ANOVA Summary}
\usage{
calc_aov_summary(aov_model)
}
\arguments{
\item{aov_model}{The ANOVA model specified via \code{stats::aov()}.}
}
\description{
Calculate ANOVA Summary
}
