% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_es_cohens_d.R
\name{calc_es_d}
\alias{calc_es_d}
\title{Calculate Effect Size Cohen's d}
\usage{
calc_es_d(
  g1_m,
  g2_m,
  g1_sd,
  g2_sd,
  g1_n,
  g2_n,
  sd_pooled = NULL,
  es_type = "hedges",
  g_correction = "standard",
  conf_level = 0.95
)
}
\arguments{
\item{g1_m}{Numeric value of group 1's mean.}

\item{g2_m}{Numeric value of group 2's mean.}

\item{g1_sd}{Numeric value of group 1's standard deviation.}

\item{g2_sd}{Numeric value of group 2's standard deviation.}

\item{g1_n}{Numeric value of group 1's sample size n.}

\item{g2_n}{Numeric value of group 2's sample size n.}

\item{sd_pooled}{Optional numeric value of the pooled standard deviation. If
not specified, the pooled standard deviation is calculated from group 1 and
2's standard deviations.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{g_correction}{Character indicating which Hedges' g correction to apply.
Options include \code{"standard"} for regular Hedges' g correction and
\code{"log"} to apply logarithmic Hedges' g correction. The default is
standard.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}
}
\description{
Calculate the Cohen's d, Hedges' g, or Glass's delta effect size
for independent samples t-tests.
}
