% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_es_means.R
\name{calc_esm}
\alias{calc_esm}
\title{Calculate Effect Size Mean Difference}
\usage{
calc_esm(
  mean,
  sd = NULL,
  n,
  sd_pooled = NULL,
  r = NULL,
  paired = FALSE,
  conf_level = 0.95,
  type = "hedges"
)
}
\arguments{
\item{mean}{}

\item{sd}{}

\item{n}{}

\item{sd_pooled}{}

\item{r}{}

\item{paired}{}

\item{conf_level}{}

\item{type}{}
}
\description{
Calculate Effect Size Mean Difference
}
