% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_anova_repeatedmeasures.R
\name{calc_rm_anova}
\alias{calc_rm_anova}
\title{Repeated Measures Analysis of Variance (ANOVA) Wrapper}
\usage{
calc_rm_anova(
  data,
  rm = "within",
  value = "value",
  id,
  within_vars,
  between_vars = NULL,
  covariates = NULL,
  interactions = FALSE,
  correction = "auto",
  posthoc_correction = "tukey",
  es_type = "hedges",
  conf_level = 0.95,
  emm_weights = TRUE
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{rm}{A character value indicating the key (i.e., name) of the repeated
measure variables.}

\item{value}{An optional character to specify the name of the scores of the
repeated measures.}

\item{id}{A required character indicating the observation ID variable when
running repeated-measures ANOVA.}

\item{within_vars}{Required characters indicating the within-subjects or
repeated variables when running repeated-measures ANOVA.}

\item{between_vars}{Optional character(s) indicating the factor grouping
variable(s).}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{interactions}{Either a logical or vector of characters representing the
interactions effects of interest. If specified as \code{TRUE}, all possible
interactionss between the groups are run (not recommended). If specified as
\code{FALSE}, no interactions terms are run. If specified with interactions
terms, e.g., \code{c("treatment\*gender", "treatment\*age_group",
"treatment\*gender\*age_group")}, the only the specified interactions terms
are run. The default is \code{FALSE}.}

\item{correction}{An optional character indicating what sphericity correction
to use. Choose between "none", "tukey", "scheffe", "bonf", and "holm"}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{conf_level}{}
}
\description{
A wrapper for running Repeated Measures Analysis of Variance
  (ANOVA).
}
