% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wilcoxon_test.R
\name{calc_wilcox_test}
\alias{calc_wilcox_test}
\title{Calculate Wilcoxon Rank-Sum / Mann-Whitney U Test}
\usage{
calc_wilcox_test(
  outcome,
  group,
  data,
  alternative = "two.sided",
  conf_level = 0.95
)
}
\arguments{
\item{outcome}{Character indicating the outcome/dependent variable.}

\item{group}{Character indicating the binary factor grouping variable.}

\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{alternative}{Character indicating the direction of the alternative
hypothesis. Options include \code{"two.sided"}, \code{"greater"}, or
\code{"less"}. The default is two sided alternative hypothesis.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}
}
\description{
Calculate Wilcoxon Rank-Sum / Mann-Whitney U independent samples t-test.
}
