% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{check_required_package}
\alias{check_required_package}
\title{Check If Required Package Is Installed}
\usage{
check_required_package(package)
}
\arguments{
\item{package}{The name of the required package.}
}
\value{
Returns TRUE or FALSE.
}
\description{
Check whether a certain package is required for a task. This includes the packages
listed under suggestions in the description file. The user will be asked
whether a package should be installed or not.
}
