% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_statistics.R
\name{code_statistics}
\alias{code_statistics}
\title{Analyze R Scripts And Print Out Statistics}
\usage{
code_statistics(paths, recursive = FALSE, output_per = "overall")
}
\arguments{
\item{paths}{A full folder path in which there are 'R' scripts to be analyzed. Can
be a single path or a vector of paths.}

\item{recursive}{FALSE by default. If TRUE scans the provided paths recursive
for sub folders and the 'R' scripts within them.}

\item{output_per}{"overall" by default. Determines how detailed the results are
printed. Valid options are "overall", "folder" or "file".}
}
\value{
Returns the results as a data frame.
}
\description{
\code{\link[=code_statistics]{code_statistics()}} reads in a folder or entire folder structure, grabs all
'R' script files and scans the contents for different patterns. It then outputs
a small report which shows of which parts the code consists.
}
\examples{
# Run this function on a directory, actually containing R script files
# to see some results.
code_statistics(tempdir())

}
