% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_report.R
\name{content_report}
\alias{content_report}
\title{Get Detailed Summary About A Data Frame}
\usage{
content_report(data_frame, output = "console", monitor = FALSE)
}
\arguments{
\item{data_frame}{The data frame to get the cointent information from.}

\item{output}{The following output formats are available: console (default) or text.}

\item{monitor}{FALSE by default. If TRUE, outputs two charts to visualize the
functions time consumption.}
}
\value{
Returns a list containing the global information as well as a data table containing
the per variable information.
}
\description{
Prints a summary of a data frames contents, including details such as variable names,
types, unique values, missings and min/max values. It also tells you the number of
observations and variables present in the data frame, memory usage and the number
of duplicate observations.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

content_report(my_data)

}
