% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabulation.R
\name{crosstab}
\alias{crosstab}
\title{Crosstabulation}
\usage{
crosstab(data = ., x, y)
}
\arguments{
\item{data}{A data frame object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{x}{A string of the first variable column name.}

\item{y}{A string of the second variable column name.}
}
\value{
A data frame with the cross-tabulation.
}
\description{
This function creates a cross-tabulation of two variables in a
  data frame, adds a total column and row, and replaces NA values with
  "<NA>".
}
\examples{
\dontrun{
crosstab(data = df, x = "var1", y = "var2")
}

}
