% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_design.R
\name{design_graphic}
\alias{design_graphic}
\title{Generate Graphics While Fully Concentrating On Design}
\usage{
design_graphic(
  data_frame,
  axes_variables = "",
  segments,
  values,
  statistics = "pct_group",
  pct_value = list(),
  compute = list(),
  formats = list(),
  by = c(),
  weight = NULL,
  titles = .qol_options[["titles"]],
  footnotes = .qol_options[["footnotes"]],
  var_labels = .qol_options[["var_labels"]],
  stat_labels = .qol_options[["stat_labels"]],
  diagram = dg_vertical_bars,
  stacked = FALSE,
  visuals = .qol_options[["graphic_visuals"]],
  axes = .qol_options[["graphic_axes"]],
  dimensions = .qol_options[["graphic_dimensions"]],
  fine_tuning = .qol_options[["graphic_fine_tuning"]],
  add_texts = NULL,
  add_forms = NULL,
  output = .qol_options[["graphic_output"]],
  na.rm = .qol_options[["na.rm"]],
  print = .qol_options[["print"]],
  monitor = .qol_options[["monitor"]]
)
}
\arguments{
\item{data_frame}{A data frame in which are the variables to tabulate.}

\item{axes_variables}{A single variable combination or a vector with single variable
that should appear on the variable axes. To nest variables use the form:
"var1 + var2 + var3 + ...".}

\item{segments}{A vector that provides variables that should appear as the graphics
segments.}

\item{values}{A vector containing all variables that should be summarised.}

\item{statistics}{Available functions:
\itemize{
\item "sum"       -> Weighted and unweighted sum
\item "sum_wgt"   -> Sum of all weights
\item "freq"      -> Unweighted frequency
\item "freq_g0"   -> Unweighted frequency of all values greater than zero
\item "pct_group" -> Weighted and unweighted percentages within the respective group
\item "pct_value" -> Weighted and unweighted percentages between value variables
\item "pct_total" -> Weighted and unweighted percentages compared to the grand total
\item "mean"      -> Weighted and unweighted mean
\item "median"    -> Weighted and unweighted median
\item "mode"      -> Weighted and unweighted mode
\item "min"       -> Minimum
\item "max"       -> Maximum
\item "sd"        -> Weighted and unweighted standard deviation
\item "variance"  -> Weighted and unweighted standard variance
\item "first"     -> First value
\item "last"      -> Last value
\item "pn"        -> Weighted and unweighted percentiles (any p1, p2, p3, ... possible)
Uses the SAS default way of calculating percentiles, which is Type 2 described in \code{\link[=quantile]{quantile()}}.
\item "missing"   -> Missings generated by the value variables
}}

\item{pct_value}{If pct_value is specified in the statistics, you can pass a list here
which contains the information for a new variable name and between which of the value
variables percentages should be computed.}

\item{compute}{A list in which individual calculations are handled. For the writing style
see the \code{\link[=compute.]{compute.()}} function.}

\item{formats}{A list in which is specified which formats should be applied to which variables.}

\item{by}{Compute tables stratified by the expressions of the provided variables.}

\item{weight}{Put in a weight variable to compute weighted results.}

\item{titles}{Specify one or more table titles.}

\item{footnotes}{Specify one or more table footnotes.}

\item{var_labels}{A list in which is specified which label should be printed for
which variable instead of the variable name.}

\item{stat_labels}{A list in which is specified which label should be printed for
which statistic instead of the statistic name.}

\item{diagram}{Function that provides the generation of the main diagram area. Built in are:
dg_vertical_bars.}

\item{stacked}{FALSE by default. If TRUE, the segments are stacked instead of grouped.
Only possible for certain diagram types.}

\item{visuals}{Visual parameters set with \code{\link[=graphic_visuals]{graphic_visuals()}}.}

\item{axes}{Axes parameters set with \code{\link[=graphic_axes]{graphic_axes()}}.}

\item{dimensions}{Dimension parameters set with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.}

\item{fine_tuning}{Fine tuning parameters set with \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}.}

\item{add_texts}{Use the \code{\link[=add_textbox]{add_textbox()}} function to freely place one or multiple textboxes.}

\item{add_forms}{Draws forms freely onto the canvas. Add forms with: \code{\link[=add_line]{add_line()}}.}

\item{output}{Output parameters set with \code{\link[=graphic_output]{graphic_output()}}.}

\item{na.rm}{FALSE by default. If TRUE removes all NA values from the variables.}

\item{print}{TRUE by default. If TRUE prints the output, if FALSE doesn't print anything.
Can be used if one only wants to catch the output data frame and workbook with meta information.}

\item{monitor}{FALSE by default. If TRUE, outputs two charts to visualize the functions
time consumption.}

\item{pct_group}{If pct_group is specified in the statistics, this option is used to
determine which variable of the row and column variables should add up to 100 \%.
Multiple variables can be specified in a vector to generate multiple group percentages.}
}
\value{
Returns a list with the data table containing the results for the graphic and the
final grid::gTree graphics object.
}
\description{
\code{\link[=design_graphic]{design_graphic()}} produces fully styled diagrams. The many parameters are all
focused around making visual changes simple and fast. All diagrams can be exported
into multiple formats.When the \href{https://CRAN.R-project.org/package=gridSVG}{gridSVG}
package is installed, it is also possible to export interactive diagrams.

The diagram parameter takes in a function for the main diagram area, making the
function modular and extendable.
}
\details{
\code{\link[=design_graphic]{design_graphic()}} is actually not based on any 'SAS' function. It is envisioned
to generate multiple different graphic iterations in no time. From a user perspective
one shouldn't have to think about "Where is the button to set the option x to y?", "In
which place do I have to put which piece of code to eventually make x happen?" or "In
which order to I have to draw which things so that hopefully the right graphic
prints on screen?". As a user you should only worry about "Do I use vertical or
horizontal bars?", "Do the colors have enough contrast?", "Do I use a legend or
direct segment labels?", "Should values be displayed or not?", "Should the graphic
be static or interactive?".

The user shouldn't be a graphics engineer, but rather a graphics designer. So the
function \code{\link[=design_graphic]{design_graphic()}} takes care of all the tedious work and only leaves you
with the fun part.
}
\examples{
# Example data frame
my_data <- dummy_data(100, insert_na = FALSE)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 65" = 18:64,
    "65 and older"   = 65:100)

age2. <- discrete_format(
    "under 18"       = 0:17,
    "18 to under 25" = 18:24,
    "25 to under 55" = 25:54,
    "55 to under 65" = 55:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Male"   = 1,
    "Female" = 2)

# Design grouped vertical bar chart
set_titles("The qol-package brings powerful concepts from 'SAS' to 'R'.")
set_footnotes("To make life easier and produce bigger",
              "and more complex outputs in less time",
              "with less code.")

my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use multiple axes grouping by just adding variables with a +
my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use multiple different segment variables within the same grouping
my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex + education",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Use different stylings. With the set_graphic_options() function the styling
# can be set globally.
set_graphic_options(primary_axes_max    = 100,
                    primary_axes_suffix = " \%",
                    color_theme         = "violet_fire",
                    title_font_color    = "#FF00FF")

my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.))

# Another possibility is to directly plug the styling into the corresponding
# function parameters.
my_data |>
     design_graphic(axes_variables = "age + education",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.),
                    dimensions = graphic_dimensions(graphic_width  = 18,
                                                    graphic_height = 12,
                                                    margins        = 0.4))

# To make a stacked diagram instead of a grouped one, use the "stacked" parameter
my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    stacked        = TRUE,
                    formats        = list(sex = sex.,
                                          age = age2.))

}
\seealso{
Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_options]{modify_graphic_options()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}

Color themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}, \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}, \code{\link[=override_theme]{override_theme()}}

Textboxes: \code{\link[=add_textbox]{add_textbox()}}, \code{\link[=get_text_width]{get_text_width()}}, \code{\link[=get_text_height]{get_text_height()}}

Viewport: \code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}, \code{\link[=back_to_the_root]{back_to_the_root()}}

Diagram: \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}, \code{\link[=setup_interactive_elements]{setup_interactive_elements()}}

Axes: \code{\link[=setup_y_axes]{setup_y_axes()}}, \code{\link[=setup_x_axes]{setup_x_axes()}}, \code{\link[=setup_xy_axes]{setup_xy_axes()}}

Labels: \code{\link[=direct_vertical_labels]{direct_vertical_labels()}}, \code{\link[=setup_legend]{setup_legend()}}

Creating formats: \code{\link[=discrete_format]{discrete_format()}} and \code{\link[=interval_format]{interval_format()}}.

Functions that can handle formats and styles: \code{\link[=frequencies]{frequencies()}}, \code{\link[=crosstabs]{crosstabs()}}, \code{\link[=any_table]{any_table()}}.

Additional functions that can handle styles: \code{\link[=export_with_style]{export_with_style()}}

Additional functions that can handle formats: \code{\link[=summarise_plus]{summarise_plus()}}, \code{\link[=recode]{recode()}},
\code{\link[=recode_multi]{recode_multi()}}, \code{\link[=transpose_plus]{transpose_plus()}}, \code{\link[=sort_plus]{sort_plus()}}
}
