% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R, R/graphic_main_grobs.R
\name{diagram}
\alias{diagram}
\alias{get_diagram_dimensions}
\alias{setup_interactive_elements}
\alias{vbar_grob}
\title{Get The Dimensions And Positions Of Groups And Segments}
\usage{
get_diagram_dimensions(arguments)

setup_interactive_elements(arguments, diagram_info)

vbar_grob(diagram_info, arguments)
}
\arguments{
\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}
}
\value{
Returns a list of parameters for groups, segments and axes.

Returns a grid::rectGrob object.

Returns a grid::rectGrob object.
}
\description{
\code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}: Get various parameters concerning the division and
position of groups and segments. These are e.g. number of groups/segments, width of
groups/segments, ids or positions among other things.

\code{\link[=vbar_grob]{vbar_grob()}}: Set up the main segments for vertical bars.

\code{\link[=vbar_grob]{vbar_grob()}}: Set up the main segments for vertical bars.
}
