% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{dimensions}
\alias{dimensions}
\alias{get_available_width}
\alias{get_available_height}
\title{Get The Available Width Or Height Without Margins}
\usage{
get_available_width(
  dimensions = .qol_options[["graphic_dimensions"]],
  total_width = NULL,
  margins = 0
)

get_available_height(
  dimensions = .qol_options[["graphic_dimensions"]],
  total_height = NULL,
  margins = 0
)
}
\arguments{
\item{dimensions}{qol package dimensions options.}

\item{total_width}{The width of the entire graphic.}

\item{margins}{The margins used on both ends of the graphic.}

\item{total_height}{The height of the entire graphic.}
}
\value{
\code{\link[=get_available_width]{get_available_width()}}: Returns a numeric value for the available width.

\code{\link[=get_available_height]{get_available_height()}}: Returns a numeric value for the available height.
}
\description{
\code{\link[=get_available_width]{get_available_width()}}: Calculate the available width from the total graphic
width minus the margins on both sides. The function can either be used with the
dimensions options from this package or with manually set width and margins.
This is an either/or option.

\code{\link[=get_available_height]{get_available_height()}}: Calculate the available height from the total graphic
height minus the margins on both sides. The function can either be used with the
dimensions options from this package or with manually set height and margins.
This is an either/or option.
}
