% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool_double_center_matrix.R
\name{double_center}
\alias{double_center}
\title{Calculate a Double-Center Matrix}
\usage{
double_center(data = ., vars, group, as_matrix = FALSE)
}
\arguments{
\item{data}{A dataframe object. A structured dataset where each column
represents a variable and each row represents an observation.}

\item{vars}{A required vector of variable names.}

\item{group}{A string character indicating the name of the group variable.}

\item{as_matrix}{A logical indicating whether to return a matrix or data
frame. Options include \code{FALSE} to return a class \code{data.frame} and
\code{TRUE} to return a class \code{matrix}.}
}
\description{
Calculate a double-centered matrix.
}
