% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{format_diagram_values}
\alias{format_diagram_values}
\title{Format Values}
\usage{
format_diagram_values(diagram_info, arguments, which = "primary")
}
\arguments{
\item{diagram_info}{The list of measurements generated by
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{which}{Primary or secondary axes.}
}
\value{
Returns a formatted value as character.
}
\description{
\code{\link[=format_diagram_values]{format_diagram_values()}}: A wrapper to make the code for formatting segment
values shorter.
}
