% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{format_values}
\alias{format_values}
\alias{format_diagram_values}
\title{Format Values}
\usage{
format_values(
  values,
  decimals = .qol_options[["graphic_axes"]][["primary_axes_decimals"]],
  big_mark = .qol_options[["graphic_axes"]][["primary_axes_big_mark"]],
  decimal_mark = .qol_options[["graphic_axes"]][["primary_axes_decimal_mark"]],
  prefix = .qol_options[["graphic_axes"]][["primary_axes_prefix"]],
  suffix = .qol_options[["graphic_axes"]][["primary_axes_suffix"]],
  scale = .qol_options[["graphic_axes"]][["primary_axes_scale"]]
)

format_diagram_values(diagram_info, arguments, which = "primary")
}
\arguments{
\item{values}{Single value or vector of values.}

\item{decimals}{Number of decimals to be displayed.}

\item{big_mark}{Character used as the thousand separator.}

\item{decimal_mark}{Character used as the decimal separator.}

\item{prefix}{What to put in front of the value.}

\item{suffix}{What to put after the value.}

\item{scale}{Multiplier for values.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{which}{Primary or secondary axes.}
}
\value{
Returns a formatted value as character.

Returns a formatted value as character.
}
\description{
\code{\link[=format_values]{format_values()}}: Format values to output numbers in a certain way.

\code{\link[=format_diagram_values]{format_diagram_values()}}: A wrapper to make the code for formatting segment
values shorter.
}
