% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ttest_report.R
\name{generate_ttest_report}
\alias{generate_ttest_report}
\title{Generate T-Test Reportable Table Sheet}
\usage{
generate_ttest_report(
  ttest_object,
  report_variance = "sd",
  report_test = "welch"
)
}
\arguments{
\item{ttest_object}{The t-test object from \code{wrap_ttest}.}

\item{report_variance}{A string value or vector of strings indicating which
variance measure to report. Options include \code{"sd"} for standard
deviation, \code{"se"} for standard error, or \code{c("sd", "se")} for
both. The default is standard deviation \code{"sd"}.}

\item{report_test}{A string value indicating which t-test statistic to
report. Options include \code{"student"} for Student's t-test,
\code{"welch"} for Welch's t-test, and \code{"wilcox"} for Wilcoxon
Rank-Sum W test / Mann-Whitney U test. The default is \code{"welch"}.}
}
\description{
Generate T-Test Reportable Table Sheet
}
