% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{get_integer_length}
\alias{get_integer_length}
\title{Get Integer Length}
\usage{
get_integer_length(variable)
}
\arguments{
\item{variable}{The integer variable from which to get the length.}
}
\value{
Returns a vector with the number of digits places.
}
\description{
Get the number of digits of an integer variable.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

my_data[["age_length"]] <- get_integer_length(my_data[["age"]])

}
