% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_pages_rmarkdown_autoupdate.R
\name{github_pages_rmarkdown_autoupdate}
\alias{github_pages_rmarkdown_autoupdate}
\title{Rmarkdown github pages}
\usage{
github_pages_rmarkdown_autoupdate(out_dir = "")
}
\value{
Returns a string that contains code.
}
\description{
Make some markdown so that when you knit an `Rmarkdown` file, it will 
automatically also make a corresponding `index.html`. This is used for when you want to 
use `github` pages.
}
\examples{
\dontrun{
github_pages_rmarkdown_autoupdate()
}

}
