% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_dimensions}
\alias{graphic_dimensions}
\title{Graphic Dimensions}
\usage{
graphic_dimensions(
  graphic_width = 16,
  graphic_height = 9,
  diagram_start_top = "auto",
  diagram_start_left = "auto",
  diagram_width = "auto",
  diagram_height = "auto",
  margins = 0.25,
  title_font_size = 9,
  footnote_font_size = 8,
  axes_font_size = 9,
  value_font_size = 9,
  label_font_size = 9,
  origin_font_size = 8,
  other_font_size = 9,
  tooltip_font_size = 12,
  line_height = 1.1,
  space_between_bars = 0,
  bar_overlap = 0,
  line_thickness = 1,
  segment_line_thickness = 1,
  major_separation_line_thickness = 1,
  minor_separation_line_thickness = 1,
  axes_line_thickness = 1,
  guiding_line_thickness = 1,
  graphic_outline_thickness = 1,
  diagram_outline_thickness = 1,
  segment_line_length = 1,
  segment_line_offset = 0,
  textbox_width = 4
)
}
\arguments{
\item{graphic_width}{The width of the whole graphic.}

\item{graphic_height}{The height of the whole graphic.}

\item{diagram_start_top}{The starting position of the main diagram within the graphic from the top.}

\item{diagram_start_left}{The starting position of the main diagram within the graphic from the left.}

\item{diagram_width}{The width of the main diagram within the graphic.}

\item{diagram_height}{The height of the main diagram within the graphic.}

\item{margins}{Inner margins to the graphic borders in cm.}

\item{title_font_size}{Font size of the title}

\item{footnote_font_size}{Font size of the footnote}

\item{axes_font_size}{Font size of the axes}

\item{value_font_size}{Font size of the values.}

\item{label_font_size}{Font size of the origin text.}

\item{origin_font_size}{Font size of the origin text.}

\item{other_font_size}{Font size of every other text element.}

\item{tooltip_font_size}{Font size of the tooltips in interactive charts.}

\item{line_height}{The height of a single text line.}

\item{space_between_bars}{The space between adjacent bars.}

\item{bar_overlap}{The factor determines how much bars are overlapping each other.}

\item{line_thickness}{The thickness of lines in a line chart or outlines of segments.}

\item{segment_line_thickness}{The thickness of the lines connecting segments with
labels.}

\item{major_separation_line_thickness}{The thickness of the lines drawn between top group
categories.}

\item{minor_separation_line_thickness}{The thickness of the lines drawn between minor group
categories.}

\item{axes_line_thickness}{The thickness of the axes lines and ticks.}

\item{guiding_line_thickness}{The thickness of the axes guiding lines.}

\item{graphic_outline_thickness}{The thickness of the graphic outline.}

\item{diagram_outline_thickness}{The thickness of the diagram outline.}

\item{segment_line_length}{The length of the lines leading from segments to labels in cm.}

\item{segment_line_offset}{Offset in the height of the leading lines in cm. If 0, all leading lines
will end at the same height, meaning the labels will be drawn at the same height. If
an offset is set, all lines after the first one will be shorter by this amount * number
of the line. Enables the labels to be drawn in steps, if e.g. the labels have a long
text which would overlap, if they where on the same height.}

\item{textbox_width}{Determines the width of any freely placed textbox.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the dimensions of a graphic produced by \code{\link[=design_graphic]{design_graphic()}}.
All parameters can also be set globally via \code{\link[=set_graphic_options]{set_graphic_options()}}.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# For default values
default_dimensions <- graphic_dimensions()

# Set specific options, the rest will be set to default values
custom_dimensions <- graphic_dimensions(graphic_width  = 10,
                                        graphic_height = 10)

# Apply options locally
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    dimensions     = custom_dimensions)

# Or direct
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    dimensions     = graphic_dimensions(graphic_width  = 10,
                                                        graphic_height = 10))

# Set options globally
set_graphic_options(graphic_width  = 10,
                    graphic_height = 10)

my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_options]{modify_graphic_options()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}

Color themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}, \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}, \code{\link[=override_theme]{override_theme()}}
}
