% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{graphic_options}
\alias{graphic_options}
\alias{set_graphic_options}
\alias{get_graphic_options}
\alias{reset_graphic_options}
\title{Set Global Graphic Options}
\usage{
set_graphic_options(..., save_file = NULL)

get_graphic_options(from_file = NULL)

reset_graphic_options()
}
\arguments{
\item{...}{Put in any graphic option from \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}},
\code{\link[=graphic_dimensions]{graphic_dimensions()}}, \code{\link[=graphic_output]{graphic_output()}} or \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}
with the new value.}

\item{save_file}{A full file path to an RDS file in which global graphic options
should be stored.}

\item{from_file}{A full file path to an RDS file in which global graphic options
are stored.}
}
\value{
\code{\link[=set_graphic_options]{set_graphic_options()}}: Returns modified global graphic options.

\code{\link[=get_graphic_options]{get_graphic_options()}}: List of global graphic options.

\code{\link[=reset_graphic_options]{reset_graphic_options()}}: Returns default global graphic options.
}
\description{
Modify styling options used by \code{\link[=design_graphic]{design_graphic()}}. Available parameters can be seen in
\code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}} and \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}.

\code{\link[=set_graphic_options]{set_graphic_options()}} sets the graphic options globally. These
options are used by all functions involved in graphics production.

\code{\link[=get_graphic_options]{get_graphic_options()}} Prints out the currently set global graphic options.

\code{\link[=reset_graphic_options]{reset_graphic_options()}} Resets global graphic options to the default parameters.
This includes all options set with \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=set_labels]{set_labels()}},
\code{\link[=set_titles]{set_titles()}} and \code{\link[=set_footnotes]{set_footnotes()}}.
}
\examples{
# This function can process any parameter from graphic_visuals(), graphic_axes(),
# graphic_dimensions(), graphic_output() or graphic_fine_tuning() and sets the
# option globally.
set_graphic_options(title_font_size = 12,
                    label_font_size = 10)

get_graphic_options()

reset_graphic_options()

}
