% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_output}
\alias{graphic_output}
\title{Graphic Output}
\usage{
graphic_output(
  save_path = NULL,
  file = NULL,
  resolution = 300,
  by_as_grid = FALSE,
  grid_columns = "auto",
  interactive = FALSE
)
}
\arguments{
\item{save_path}{If NULL, only draws the graphic in the plot window. Otherwise
specify an output path.}

\item{file}{If NULL, only draws the graphic in the plot window. Otherwise specify
a filename with extension.}

\item{resolution}{DPI resolution.}

\item{by_as_grid}{FALSE by default. If TRUE exports one single graphic containing the
whole grid when using by variables. If used with interactive parameter all graphics
in the grid will be put into one file, where the graphics can be switched with a
drop down list.}

\item{grid_columns}{By default tries to create a square shape. If a numeric value is
passed the number of grid columns is fixed and the number of rows adjusts accordingly.}

\item{interactive}{FALSE by default. If TRUE exports an interactive html file containing
mouse over tooltips and hover effects.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the dimensions of a graphic produced by \code{\link[=design_graphic]{design_graphic()}}.
All parameters can also be set globally via \code{\link[=set_graphic_options]{set_graphic_options()}}.
}
\details{
\if{html}{\out{<div class="sourceCode html">}}\preformatted{Interactive graphics can be embedded in an <iframe> like this:

<iframe src="my_graphic.html" style="width: 100\%; aspect-ratio: 16 / 9; border: none;"
scrolling="no"></iframe>
}\if{html}{\out{</div>}}

Set the aspect ratio according to your graphic dimensions.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# For default values
default_output <- graphic_output()

# Set specific options, the rest will be set to default values
custom_output <- graphic_output(save_path = "C:/MyFolder/",
                                file      = "MyGraphic.png")

# Apply options locally
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    output         = custom_output)

# Or direct
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    output         = graphic_output(save_path = "C:/MyFolder/",
                                                    file      = "MyGraphic.png"))

# Set options globally
set_graphic_options(save_path = "C:/MyFolder/",
                    file      = "MyGraphic.png")

my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_options]{modify_graphic_options()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}

Color themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}, \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}, \code{\link[=override_theme]{override_theme()}}
}
