% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepl_get.R
\name{grepl_get}
\alias{grepl_get}
\title{grepl names}
\usage{
grepl_get(pattern, x, x_names = T, negate = FALSE, ignore.case = FALSE)
}
\arguments{
\item{pattern}{character string containing a regular experssion.}

\item{x}{An indexable object with optional names.}

\item{x_names}{logical. If \code{TRUE}, return matching names of \code{x}. Otherwise,
return the matching elements of \code{x}.}

\item{negate}{logical. If \code{TRUE}, negates the matching of \code{pattern}.}

\item{ignore.case}{if \code{FALSE}, the pattern matching is case sensitive.}
}
\value{
Returns a subset of \code{x} or the names of \code{x} that are matched to the
supplied regex.
}
\description{
grepl names
}
