% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{hex_ansi}
\alias{hex_ansi}
\alias{hex_to_256}
\alias{hex_to_ansi}
\title{Convert Color Codes}
\usage{
hex_to_256(hex_color)

hex_to_ansi(text, hex_color = NULL, bold = FALSE)
}
\arguments{
\item{hex_color}{The hex color code to convert.}

\item{text}{The text which contains hex color codes to be convert into ansi
style formatting.}

\item{bold}{FALSE by default. If TRUE inserts ansi code for bold printing}
}
\value{
\code{\link[=hex_to_256]{hex_to_256()}}: Returns 6x6x6 color cube.

\code{\link[=hex_to_ansi]{hex_to_ansi()}}: Returns formatted text.
}
\description{
\code{\link[=hex_to_256]{hex_to_256()}}: Generate a 256-color 6x6x6 color cube.

\code{\link[=hex_to_ansi]{hex_to_ansi()}}: Applies hex color and font weight to a text as ansi codes.
}
\examples{
color_cube <- hex_to_ansi("#32CD32")

formatted_text <- hex_to_ansi("This is a text to test the coloring",
                              hex_color = "#32CD32", bold = TRUE)

}
