% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_random_noise.R
\name{inject_noise}
\alias{inject_noise}
\title{Inject Random Noise}
\usage{
inject_noise(data = ., iter = 1, vars = NULL, noise_value = 1, digits = NA)
}
\arguments{
\item{data}{}

\item{iter}{}

\item{vars}{}
}
\description{
Inject Random Noise
}
