% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_discriminant_function_analysis.R
\name{ldfa}
\alias{ldfa}
\title{Linear Discriminant Analysis Function Coefficients}
\usage{
ldfa(data, outcome, inputs)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcome}{A required string of a column name representing the outcome
variable.}

\item{inputs}{A required vector of column names representing the input
variables.}
}
\value{
The standard \code{lda} model object with additional appended
  coefficients and predicted classes within the list.
}
\description{
Runs linear discriminant analysis (LDA) through the \code{MASS}
  package's \code{lda()} function. As the \code{lda()} function does not
  offer canonical function coefficients by default, this function calculates
  this to mirror SPSS's output of coefficients through their discriminant
  function analysis (DFA) approach.
}
\examples{
ldfa(iris,
     outcome = "species",
     inputs = c("Sepal.Length", "Sepal.Width",
                "Petal.Length", "Petal.Width"))

}
\references{
Venables, W. N. & Ripley, B. D. (2002) Modern Applied Statistics
  with S. Fourth Edition. Springer, New York. ISBN 0-387-95457-0
}
