% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{libname}
\alias{libname}
\title{Check If Path Exists And Retrieve Files}
\usage{
libname(path, get_files = FALSE)
}
\arguments{
\item{path}{A folder path.}

\item{get_files}{FALSE by default. If TRUE returns a named character vector containing file paths.}
}
\value{
Returns the given file path or a named character vector containing file paths.
}
\description{
Libname checks if a given path exists and writes a message in the console accordingly.
Optional all files from the given path can be retrieved as a named character vector.
}
\examples{
my_path   <- libname("C:/My_Path/")
file_list <- libname("C:/My_Path/", get_files = TRUE)

}
