% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_diagrams.R
\name{main_diagram}
\alias{main_diagram}
\alias{dg_vertical_bars}
\title{Generate The Main Diagram Area}
\usage{
dg_vertical_bars(arguments, diagram_info)
}
\arguments{
\item{arguments}{Argument list passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}
}
\value{
Returns a grid::gTree object.
}
\description{
The listed functions are the built in diagram types which ship with this package.
These functions don't work on their own and have to be plugged into the "diagram"
parameter of \code{\link[=design_graphic]{design_graphic()}}.

\code{\link[=dg_vertical_bars]{dg_vertical_bars()}}: Grouped vertical bars.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 65" = 18:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Total"  = 1:2,
    "Male"   = 1,
    "Female" = 2)

# Design grouped vertical bar chart
grouped_vbars <- my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "age",
                    values         = weight,
                    diagram        = dg_vertical_bars,
                    formats        = list(sex = sex.,
                                          age = age.))

}
