% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_main_grobs.R
\name{main_grobs}
\alias{main_grobs}
\alias{vertical_grouped_bar_grob}
\alias{vertical_stacked_bar_grob}
\title{Generate The Main Graphical Objects}
\usage{
vertical_grouped_bar_grob(diagram_info, arguments)

vertical_stacked_bar_grob(diagram_info, arguments)
}
\arguments{
\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}
}
\value{
Returns a grid::gList containing segments and labels.
}
\description{
These functions create the graphical objects for the main diagram area of the
different diagram types. These functions don't work on their own. They have to
be used inside a custom diagram function, which builds the whole diagram and can
be plugged into the "diagram" parameter of \code{\link[=design_graphic]{design_graphic()}}.

\code{\link[=vertical_grouped_bar_grob]{vertical_grouped_bar_grob()}}: Set up the main segments for vertical bars.

\code{\link[=vertical_stacked_bar_grob]{vertical_stacked_bar_grob()}}: Set up the main segments for vertical bars.
}
