% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{modify_graphic_axes}
\alias{modify_graphic_axes}
\title{Modify Graphic Axes}
\usage{
modify_graphic_axes(axes_to_modify, ...)
}
\arguments{
\item{axes_to_modify}{A pre created graphic visuals object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic
dimension elements.}
}
\value{
Returns a modified list of named graphic options.
}
\description{
Modify previously set up graphic visual appearance with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.
}
\examples{
# For default values
ga <- graphic_axes()

# Set specific options, the rest will be set to default values
ga <- graphic_axes(primary_axes_max      = 100,
                   primary_axes_decimals = 1)

# Modify the previously created graphics object
ga <- ga |> modify_graphic_axes(primary_axes_min = 50)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
