% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{modify_graphic_dimensions}
\alias{modify_graphic_dimensions}
\title{Modify Graphic Dimensions}
\usage{
modify_graphic_dimensions(dimension_to_modify, ...)
}
\arguments{
\item{dimension_to_modify}{A pre created graphic dimension object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic
dimension elements.}
}
\value{
Returns a modified list of named graphic options.
}
\description{
Modify previously set up graphic dimensions with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.
}
\examples{
# For default values
gd <- graphic_dimensions()

# Set specific options, the rest will be set to default values
gd <- graphic_dimensions(graphic_width  = 10,
                         graphic_height = 10)

# Modify the previously created graphics object
gd <- gd |> modify_graphic_dimensions(title_font_size = 12)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
