% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{modify_graphic_fine_tuning}
\alias{modify_graphic_fine_tuning}
\title{Modify Graphic Dimensions}
\usage{
modify_graphic_fine_tuning(fine_tune_to_modify, ...)
}
\arguments{
\item{fine_tune_to_modify}{A pre created graphic fine tuning object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic
fine tuning elements.}
}
\value{
Returns a modified list of named graphic options.
}
\description{
Modify previously set up graphic dimensions with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.
}
\examples{
# For default values
gft <- graphic_fine_tuning()

# Set specific options, the rest will be set to default values
gft <- graphic_fine_tuning(diagram_start_adjust = 2.5,
                           tick_length          = 0.05)

# Modify the previously created graphics object
gft <- gft |> modify_graphic_fine_tuning(diagram_margin = 0.02)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
