% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{modify_graphic_output}
\alias{modify_graphic_output}
\title{Modify Graphic Output}
\usage{
modify_graphic_output(output_to_modify, ...)
}
\arguments{
\item{output_to_modify}{A pre created graphic output object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic
output elements.}
}
\value{
Returns a modified list of named graphic options.
}
\description{
Modify previously set up graphic output with \code{\link[=graphic_output]{graphic_output()}}.
}
\examples{
# For default values
gout <- graphic_output()

# Set specific options, the rest will be set to default values
gout <- graphic_output(save_path = "C:/MyFolder/",
                       file      = "MyGraphic.png")

# Modify the previously created graphics object
gout <- gout |> modify_graphic_output(file = "MyGraphic.pdf")

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
