% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{modify_graphic_visuals}
\alias{modify_graphic_visuals}
\title{Modify Graphic Visuals}
\usage{
modify_graphic_visuals(visuals_to_modify, ...)
}
\arguments{
\item{visuals_to_modify}{A pre created graphic visuals object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic
dimension elements.}
}
\value{
Returns a modified list of named graphic options.
}
\description{
Modify previously set up graphic visual appearance with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.
}
\examples{
# For default values
gv <- graphic_visuals()

# Set specific options, the rest will be set to default values
gv <- graphic_visuals(title_font_color = "#00FF00",
                      line_markers     = FALSE)

# Modify the previously created graphics object
gv <- gv |> modify_graphic_visuals(label_group = 2)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
