% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_explorer.R
\name{open_explorer}
\alias{open_explorer}
\title{Open explorer}
\usage{
open_explorer(path = NULL)
}
\arguments{
\item{path}{}
}
\value{
No return values. This function is invoked for the side-effect of opening Windows 
explorer at a given directory.
}
\description{
Open explorer
}
\examples{
\dontrun{
# Open Windows explorer in the current directory
open_explorer()
# Open windows explorer in your desktop
open_explorer("C:/whoami/Desktop")
}

}
