% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_least_squares_sem_optimizer.R
\name{plssem_optimizer}
\alias{plssem_optimizer}
\title{Partial Least Squares Structural Equation Modeling Wrapper}
\usage{
plssem_optimizer(
  loadings = NULL,
  weights = NULL,
  data,
  total_estimates,
  outcomes,
  weight_type = "combo",
  cor_method = "pearson"
)
}
\arguments{
\item{loadings}{}

\item{weights}{}

\item{data}{}

\item{total_estimates}{}

\item{outcomes}{}

\item{weight_type}{}

\item{cor_method}{}
}
\description{
Partial Least Squares Structural Equation Modeling Wrapper
}
