% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{unit_conversion}
\alias{unit_conversion}
\alias{convert_to_native_width}
\alias{convert_to_native_height}
\title{Convert Values Into Native Units}
\usage{
convert_to_native_width(value, diagram_info, dimensions, base = "diagram")

convert_to_native_height(value, diagram_info, dimensions, base = "diagram")
}
\arguments{
\item{value}{The value to convert into native units.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{dimensions}{qol package dimensions options.}

\item{base}{"diagram" by default. Can be either "diagram" or "graphic", depending on
what part of the graphic the value should be measured.}
}
\value{
Returns numeric value scaled to native units.
}
\description{
Native units are scaled from 0 to 1 and take to individual viewport and axes
scaling into account.

\code{\link[=convert_to_native_width]{convert_to_native_width()}}: Convert a value into a native width unit.

\code{\link[=convert_to_native_height]{convert_to_native_height()}}: Convert a value into a native height unit.
}
