% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_diagrams.R
\name{vbars}
\alias{vbars}
\title{Generate The Main Graphic Area For VBars}
\usage{
vbars(...)
}
\arguments{
\item{...}{Arguments passed on by \code{\link[=design_graphic]{design_graphic()}}.}
}
\value{
Returns a grid::gTree object.
}
\description{
Generate the main graphic area for vbars
}
