% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{viewport}
\alias{viewport}
\alias{setup_nested_diagram_viewport}
\alias{get_vertical_value_fit}
\alias{back_to_the_root}
\title{Set Up Viewports To Draw On}
\usage{
setup_nested_diagram_viewport(arguments, diagram_info)

get_vertical_value_fit(diagram_info, dimensions, visuals, fine_tuning, stacked)

back_to_the_root(to_main = TRUE)
}
\arguments{
\item{arguments}{Argument list passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{dimensions}{Dimension parameters set with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.}

\item{visuals}{Visual parameters set with \code{\link[=graphic_visuals]{graphic_visuals()}}.}

\item{fine_tuning}{Fine tuning parameters set with \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}.}

\item{stacked}{FALSE by default. If TRUE, the segments are stacked instead of grouped.
Only possible for certain diagram types.}

\item{to_main}{TRUE by default. Pops out of all viewports but the main canvas.}
}
\value{
Returns the grid::current.vpPath.

Returns the grid::current.vpPath.
}
\description{
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}: Setup the main diagram viewport based on
the list of arguments passed on by \code{\link[=design_graphic]{design_graphic()}}. This viewport contains
the main diagram area as well as the axes. Additionally sets up the main diagram
area viewport which only contains the main drawing area.

\code{\link[=back_to_the_root]{back_to_the_root()}}: Pops out of all viewports.

\code{\link[=back_to_the_root]{back_to_the_root()}}: Pops out of all viewports.
}
