% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_else.R
\name{where.}
\alias{where.}
\title{Filter Data Frame With Direct View}
\usage{
where.(data_frame, condition = NULL, keep = NULL)
}
\arguments{
\item{data_frame}{A data frame on which to apply filters.}

\item{condition}{The condition on which to filter observations.}

\item{keep}{The Variables to keep in the result data frame.}
}
\value{
Returns a filtered data frame.
}
\description{
Filter observations and variables and directly view the result on screen.
}
\examples{
# Example data frame
my_data <- dummy_data(1000)

# Get a quick filtered view
my_data |> where.(sex == 1 & age < 25,
                  c(sex, age, household_id, education))

}
\seealso{
The following functions can make use of the \code{\link[=do_if]{do_if()}} filter variables:

Conditions: \code{\link[=if.]{if.()}}, \code{\link[=else_if.]{else_if.()}}, \code{\link[=else.]{else.()}}

Filter Data Frame: \code{\link[=where.]{where.()}}
}
