% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_anova.R
\name{wrap_anova}
\alias{wrap_anova}
\title{Analysis of Variance (ANOVA) Wrapper}
\usage{
wrap_anova(
  data,
  outcome = NULL,
  between_vars = NULL,
  within_vars = NULL,
  rm = "within",
  value = "value",
  id = NULL,
  covariates = NULL,
  interactions = FALSE,
  rm_correction = "auto",
  posthoc_correction = "tukey",
  conf_level = 0.95,
  es_type = "hedges",
  emm_weights = TRUE
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcome}{A required character indicating the outcome/dependent variable
when running one-way ANOVA.}

\item{between_vars}{Required character(s) indicating the factor grouping
variable(s) when running one-way ANOVA. Optional when running
repeated-measures ANOVA.}

\item{within_vars}{Required characters indicating the within-subjects or
repeated variables when running repeated-measures ANOVA.}

\item{rm}{An optional character to specify the name of the repeated
measures.}

\item{value}{An optional character to specify the name of the scores of the
repeated measures.}

\item{id}{A required character indicating the observation ID variable when
running repeated-measures ANOVA.}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{interactionss}{Either a logical or vector of characters representing the
interactions effects of interest. If specified as \code{TRUE}, all possible
interactionss between the groups are run (not recommended). If specified as
\code{FALSE}, no interactions terms are run. If specified with interactions
terms, e.g., \code{c("treatment\*gender", "treatment\*age_group",
"treatment\*gender\*age_group")}, the only the specified interactions terms
are run. The default is \code{FALSE}.}

\item{ph_method}{The post-hoc pairwise comparison correction method. Options
include \code{"tukey"}, \code{"scheffe"} , \code{"sidak"},
\code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and \code{"none"}.
The default is \code{"tukey"}. See below for additional information.}
}
\description{
Analysis of Variance (ANOVA) Wrapper
}
