% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_sem_partial_least_squares.R
\name{wrap_plssem}
\alias{wrap_plssem}
\title{Partial Least Squares Structural Equation Modeling Wrapper}
\usage{
wrap_plssem(
  data = .,
  measurements,
  structure,
  bootn = 1000,
  inner_weights = seminr::path_weighting,
  missing = seminr::mean_replacement,
  missing_value = NA,
  max_iter = 300,
  stop_criterion = 7,
  digits = 3,
  prioritization = FALSE,
  prioritization_type = "combo",
  prioritization_weight_type = "combo",
  prioritization_cor_method = "pearson"
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{measurements}{Measurement model object created using \code{constructs}
function from the \code{seminr} package. Refer to the \code{seminr} help
documentation for more details.}

\item{structure}{Structural model object created using the
\code{relationships} function from \code{seminr} package. Refer to the
\code{seminr} help documentation for more details.}

\item{bootn}{A numeric value specifying the number of bootstrap iterations to
perform. The default value is \code{1000}. If specified to \code{0}, then
no bootstrapping is performed.}

\item{inner_weights}{A \code{seminr} parameter to specify the inner weighting
scheme in the estimation of the inner paths matrix. Options include
\code{seminr::path_weighting} (default) or \code{path_factorial}.}

\item{missing}{A \code{seminr} parameter to specify how the model should
handle missing values. The default is \code{seminr::mean_replacement}.}

\item{missing_value}{A value indicating the missing value in the dataset. The
default is \code{NA}.}

\item{max_iter}{A numeric value indicating the maximum number of iterations
to run when estimating PLS-SEM. The default value is \code{300}.}

\item{stop_criterion}{A numeric value indicating the stop criterion for
estimating PLS-SEM. The default value is \code{7}.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{prioritization}{A logical value for whether to run feature
prioritization analysis.}

\item{prioritization_type}{A string value indicating which weighting scheme
to use in calculating the importance of individual features to their
composite variable. Options include \code{"loading"} to weight by indicator
composite loading, \code{"weight"} to weight by indicator composite
weights, or \code{"combo"} to weight by both indicator composite loadings
and weights. The default is \code{"combo"}.}

\item{prioritization_weight_type}{A string value indicating which weighting
scheme to use in calculating the importance of individual features at the
lower-order level. Options include \code{"se"} to weight by standard error,
\code{"est"} to weight by item loading, or \code{"combo"} to weight by both
standard error and item loading. The default is \code{"combo"}.}

\item{prioritization_cor_method}{A string value indicating which correlation
coefficient to use in calculating the halo-effect feature weighting.
Options include \code{"pearson"} for Pearson's correlation,
\code{"kendall"} for Kendall rank correlation, and \code{"spearman"} for
Spearman's rank correlation. The default is \code{"pearson"}.}

\item{file}{Location to save output file as excel workbook if specified.}
}
\description{
A wrapper function to automate PLS-SEM and extracting its
  results. The current function is limited to only 2 serial mediators. Due to
  the undue level of complexity added to the model estimations when there are
  3 or more serial mediators, the current function is not designed to handle
  this type of model. This automation function uses the \code{seminr} package
  (Ray et al., 2022).
}
\examples{
measurements <- constructs(
  composite("comp1", multi_items("com", 1:3)),
  composite("comp2", multi_items("com", 4:6)),
  higher_composite("high_comp", c("comp1", "comp2")),
  composite("mani1", c("v1")),
  composite("mani2", c("v2")),
  composite("outcome", multi_items("out", 1:5)),
  interaction_term(iv = "high_comp", moderator = "mani1", method = two_stage)
)

structure <- relationships(
  paths(from = c("high_comp"),
        to = c("mani1", "mani2")),
  paths(from = c("high_comp", "mani1", "mani2", "high_comp*mani1"),
        to = c("outcome"))
)

plssem_wrapper(data,
               measurements,
               structure)


}
\references{
Ray S, Danks N, Calero Valdez A (2022). seminr: Building and
  Estimating Structural Equation Models. R package version 2.3.2,
  \url{https://CRAN.R-project.org/package=seminr}.
}
