% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_repeatedmeasures_wrapper.R
\name{wrap_rm_anova}
\alias{wrap_rm_anova}
\title{Repeated Measures Analysis of Variance (ANOVA) Wrapper}
\usage{
wrap_rm_anova(
  data,
  key = "within",
  value = "value",
  id,
  within_vars,
  between_vars = NULL,
  covariates = NULL,
  interaction = FALSE,
  conf_level = 0.95,
  ph_method = "holm",
  es_type = "hedges"
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{key}{An optional character to specify the name of the repeated
measures.}

\item{value}{An optional character to specify the name of the scores of the
repeated measures.}

\item{within_vars}{Characters indicating the repeated measures variables.}

\item{between_vars}{An optional character indicating the between groups
factor grouping variable.}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{interaction}{Either a logical or vector of characters representing the
interaction effects of interest. If specified as \code{TRUE}, all possible
interactions between the groups are run (not recommended). If specified as
\code{FALSE}, no interaction terms are run. If specified with interaction
terms, e.g., \code{c("treatment\*gender", "treatment\*age_group",
"treatment\*gender\*age_group")}, the only the specified interaction terms
are run. The default is \code{FALSE}.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{ph_method}{The post-hoc pairwise comparison correction method. Options
include \code{"tukey"}, \code{"scheffe"} , \code{"sidak"},
\code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and \code{"none"}.
The default is \code{"tukey"}. See below for additional information.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}
}
\description{
A wrapper for running Repeated Measures Analysis of Variance
  (ANOVA).
}
