% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_ttest.R
\name{wrap_ttest}
\alias{wrap_ttest}
\title{Independent Samples T-Test Wrapper}
\usage{
wrap_ttest(
  data = .,
  outcomes,
  group,
  test = c("student", "welch", "wilcox"),
  alternative = "two.sided",
  conf_level = 0.95,
  digits = 3,
  es_type = "hedges",
  filename = NULL,
  study_name = NULL,
  assumptions = FALSE,
  report_variance = "sd",
  report_test = "welch"
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcomes}{A vector of characters indicating the outcome/dependent
variables.}

\item{group}{Character indicating the binary factor grouping variable.}

\item{test}{A vector of characters indicating which independent samples test
to run. Options include \code{"student"} for Student's t-test,
\code{"welch"} for Welch's t-test, and \code{"wilcox"} for
Wilcoxon/Mann-Whitney U test. The default is to run all three.}

\item{alternative}{Character indicating the direction of the alternative
hypothesis. Options include \code{"two.sided"}, \code{"greater"}, or
\code{"less"}. The default is two sided alternative hypothesis.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{digits}{Number of decimal places for the correlation matrix. Default is
3 decimal points. If not specified, the function will use the default
value.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{filename}{Location to save output file as excel workbook if specified.}

\item{study_name}{Relevant only if \code{filename} is specified. An optional
character string indicating the name of the study.}

\item{assumptions}{Relevant only if \code{filename} is specified. A logical
indicating whether to add an t-test assumptions sheet. The default is
\code{FALSE}.}

\item{report_variance}{Relevant only if \code{filename} is specified. A
string indicating which variance statistic to report. Options include
\code{"sd"} for standard deviation and \code{"se"} for standard error. The
default is \code{"sd"}.}

\item{report_test}{Relevant only if \code{filename} is specified. A string
logical indicating which t-test to include in the report. Options include
\code{"student"} for Student's t-test, \code{"welch"} for Welch's t-test,
and \code{"wilcoxon"} for Wilcoxon rank-sum test (i.e., Mann-Whitney U
test). The default is \code{"welch"}.}
}
\description{
A wrapper function to automate running and extracting the
  results of independent samples t-tests.
}
\details{
The current function is limited to running Student's, Welch's, and
  Wilcoxon/Mann-Whitney U tests. Binary group effect sizes are limited to
  Cohen's d, Hedges' g, and Glass's delta.
}
