% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anova_workbook.R
\name{write_anova}
\alias{write_anova}
\title{Write ANOVA Report Workbook}
\usage{
write_anova(
  anova_object,
  filename,
  study_name = NULL,
  assumptions = FALSE,
  digits = 3,
  summary_type = 1,
  report_es = NULL
)
}
\arguments{
\item{anova_object}{The ANOVA object from \code{wrap_anova}.}

\item{filename}{The filepath to save the workbook.}

\item{study_name}{An optional character string indicating the name of the
study.}

\item{assumptions}{A logical indicating whether to add an ANOVA assumptions
sheet. The default is FALSE.}

\item{digits}{A numeric value indicating the number of digits to round to.}
}
\description{
Write ANOVA Report Workbook
}
