% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_plssem_workbook.R
\name{write_plssem_workbook}
\alias{write_plssem_workbook}
\title{Write PLS-SEM Report Workbook}
\usage{
write_plssem_workbook(
  plssem_object,
  filename,
  study_name = NULL,
  digits = 3,
  path_tables = c("long", "wide")
)
}
\arguments{
\item{plssem_object}{The PLS-SEM object from \code{wrap_plssem}.}

\item{filename}{The filepath to save the workbook.}

\item{study_name}{An optional character string indicating the name of the
study.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{path_tables}{A vector of strings indicating which direct effects
pathway table to export. Options include \code{"long"} for long-formatted
regression tables with confidence intervals below the coefficients,
\code{"wide"} for wide-formatted regression tables with confidence
intervals adjacent to the coefficients, or \code{NULL} to not export any.}
}
\description{
Write PLS-SEM Report Workbook
}
