% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ttest_workbook.R
\name{write_ttest}
\alias{write_ttest}
\title{Write T-Test Report Workbook}
\usage{
write_ttest(
  ttest_object,
  filename,
  study_name = NULL,
  assumptions = FALSE,
  digits = 3,
  report_variance = "sd",
  report_test = "welch"
)
}
\arguments{
\item{ttest_object}{The t-test object from \code{wrap_ttest}.}

\item{filename}{The filepath to save the workbook.}

\item{study_name}{An optional character string indicating the name of the
study.}

\item{assumptions}{A logical indicating whether to add an t-test assumptions
sheet. The default is \code{FALSE}.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{report_variance}{A string indicating w}
}
\description{
Write T-Test Report Workbook
}
