% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_anova_report.R
\name{add_anova_report}
\alias{add_anova_report}
\title{Add ANOVA Reportable Table Format}
\usage{
add_anova_report(
  wb,
  sheet_name,
  anova_object,
  name = NULL,
  digits = 3,
  report_es = c("etaSqP")
)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{anova_object}{}

\item{name}{}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{report_es}{A vector of strings indicating which effect sizes to report.
Options include \code{"ges"} for generalized eta-squared, \code{"etaSq"}
for eta-squared, \code{"etaSqP"} for partial eta-squared, \code{"omegaSq"}
for omega-squared, and \code{"cohen_f"} for Cohen's f. Specify \code{NULL}
to not report any effect size.}
}
\description{
Add ANOVA Reportable Table Format
}
