% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ttest_report.R
\name{add_ttest_report}
\alias{add_ttest_report}
\title{Add T-Test Reportable Table Format}
\usage{
add_ttest_report(
  wb,
  sheet_name,
  ttest_object,
  name = NULL,
  digits = 3,
  report_variance,
  report_test
)
}
\arguments{
\item{wb}{}

\item{sheet_name}{}

\item{ttest_object}{}

\item{name}{}

\item{digits}{}
}
\description{
Add T-Test Reportable Table Format
}
