% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{add_variable_range}
\alias{add_variable_range}
\title{Add Empty Variables In A Given Range}
\usage{
add_variable_range(data_frame, var_range)
}
\arguments{
\item{data_frame}{A data frame to add variables to.}

\item{var_range}{A range of variables to add, provided in the form: var_name1:var_name10.}
}
\value{
Returns a data frame with added variables.
}
\description{
Add empty variables to a data frame in the provided range. Basically does in a
data frame, what paste0("age", 1:10) does for a vector.
}
\examples{
# Example data frames
my_data <- dummy_data(100)

# Add variable range
my_data <- my_data |> add_variable_range(status1:status12)

}
