% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{axes}
\alias{axes}
\alias{setup_y_axes}
\alias{setup_x_axes}
\alias{setup_xy_axes}
\alias{setup_guiding_lines}
\alias{swap_scaling}
\alias{even_prettier}
\title{Measure And Generate Custom Axes}
\usage{
setup_y_axes(diagram_info, arguments, secondary = FALSE)

setup_x_axes(diagram_info, arguments)

setup_xy_axes(diagram_info, arguments)

setup_guiding_lines(diagram_info, arguments)

swap_scaling(measuring, from = "width", to = "height")

even_prettier(min_value, max_value, number_of_steps)
}
\arguments{
\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{arguments}{Argument list passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{secondary}{FALSE by default. If TRUE draws the primary axes on the left side,
otherwise the secondary axes on the right side.}

\item{measuring}{The values received from \code{\link[=get_text_width]{get_text_width()}} or \code{\link[=get_text_height]{get_text_height()}}.}

\item{from}{Input measuring.}

\item{to}{Output measuring.}

\item{min_value}{Minimum starting value.}

\item{max_value}{Maximum end value.}

\item{number_of_steps}{The number of equal length steps to be used to head from
the minimum to the maximum value.}
}
\value{
\code{\link[=setup_y_axes]{setup_y_axes()}}, \code{\link[=setup_x_axes]{setup_x_axes()}}, \code{\link[=setup_xy_axes]{setup_xy_axes()}}: Returns a grid::gTree
object containing the full axes.

\code{\link[=swap_scaling]{swap_scaling()}}: Returns a vector of width or heights.

\code{\link[=even_prettier]{even_prettier()}}: Returns a pretty sequence of equal length numeric values.
}
\description{
\code{\link[=setup_y_axes]{setup_y_axes()}}: Uses the currently active viewport to set up the y axes. The
primary axes is drawn on the left, the secondary axes on the right side of the
currently active viewport.

\code{\link[=setup_x_axes]{setup_x_axes()}}: Uses the currently active viewport to set up the x axes. The
height at which it is drawn is determined by the primary y axes 0 point. If y0
is not on the y axes then the x axes will either be drawn at the bottom, if only
positive values are present, or at the top, if only negative values are present.

\code{\link[=setup_xy_axes]{setup_xy_axes()}}: Uses the currently active viewport to set up both the x and
y axes.

\code{\link[=setup_guiding_lines]{setup_guiding_lines()}}: Uses the currently active viewport to set up the x and
y axes guiding lines. Guiding lines are drawn at the tick positions and reach from
the axes to the opposite end of the diagram.

\code{\link[=swap_scaling]{swap_scaling()}}: Width and height are measured according to the individual scaling
for each dimension. This function converts the measured to the respective other dimension.

\code{\link[=even_prettier]{even_prettier()}}: Generates a sequence of pretty steps for equal length ticks.
The function calculates more granular results than the base \code{\link[=pretty]{pretty()}} function.
}
