% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{axes}
\alias{axes}
\alias{get_value_axes_width}
\alias{get_variable_axes_height}
\alias{get_values_width}
\alias{get_values_height}
\alias{swap_xy_scaling}
\alias{get_group_tick_positions_x}
\alias{get_y_axes_values}
\alias{setup_y_axes}
\alias{setup_x_axes}
\alias{setup_xy_axes}
\alias{inject_inner_canvas_size}
\title{Measure And Generate Custom Axes}
\usage{
get_value_axes_width(
  graphic_tab,
  max_value,
  axes,
  dimensions,
  visuals,
  fine_tuning,
  which = "primary"
)

get_variable_axes_height(wrapped_text, dimensions, visuals, fine_tuning)

get_values_width(diagram_info, dimensions, visuals, arguments)

get_values_height(diagram_info, dimensions, visuals, fine_tuning)

swap_xy_scaling(
  measuring,
  dimensions,
  from = "width",
  to = "height",
  part = "diagram"
)

get_group_tick_positions_x(
  number_of_groups,
  number_of_segments,
  segment_width,
  segment_pos
)

get_y_axes_values(values, axes, fine_tuning, which = "primary")

setup_y_axes(tick_positions, tick_values, arguments, secondary = FALSE)

setup_x_axes(tick_positions, label_positions, labels, zero_pos, arguments)

setup_xy_axes(diagram_info, arguments)

inject_inner_canvas_size(axes, arguments)
}
\arguments{
\item{graphic_tab}{The base data frame for the graphic.}

\item{max_value}{The maximum value which takes the most space.}

\item{axes}{The axes gTree generated by the setup axes functions.}

\item{dimensions}{The list of dimensions parameters.}

\item{visuals}{The list of visual parameters.}

\item{fine_tuning}{The list of fine tuning parameters.}

\item{which}{Primary or secondary axes.}

\item{wrapped_text}{The wrapped text for the variable axes to measure.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{measuring}{The values received from \code{\link[=get_values_width]{get_values_width()}} or \code{\link[=get_values_height]{get_values_height()}}.}

\item{from}{Input measuring: Can be "width" or "height".}

\item{to}{Output measuring: Can be "width" or "height".}

\item{part}{Determine measuring on which part of the graphic. Can be "graphic", "diagram" or
"inner_canvas".}

\item{number_of_groups}{The number of groups.}

\item{number_of_segments}{The number of segments per group.}

\item{segment_width}{The width of a single segment}

\item{segment_pos}{A vector containing all segment positions.}

\item{values}{Value vector from the data frame.}

\item{tick_positions}{The position of the axes ticks.}

\item{tick_values}{The values drawn at the tick positions.}

\item{secondary}{FALSE by default. If TRUE draws the primary axes on the left side,
otherwise the secondary axes on the right side.}

\item{label_positions}{The positions the labels will be drawn at.}

\item{labels}{The labels per tick or per segment group.}

\item{zero_pos}{The y axes 0 position at which the x axes is drawn.}
}
\value{
\code{\link[=get_value_axes_width]{get_value_axes_width()}}: Returns a numeric width.

\code{\link[=get_variable_axes_height]{get_variable_axes_height()}}: Returns a numeric height.

\code{\link[=get_values_width]{get_values_width()}}: Returns a numeric width

\code{\link[=get_values_height]{get_values_height()}}: Returns a numeric height.

\code{\link[=swap_xy_scaling]{swap_xy_scaling()}}: Returns a vector of width or heights.

\code{\link[=get_group_tick_positions_x]{get_group_tick_positions_x()}}: Returns a vector of numeric tick positions.

\code{\link[=get_y_axes_values]{get_y_axes_values()}}: Returns a numeric vector of equally spaced round values.

\code{\link[=setup_y_axes]{setup_y_axes()}}, \code{\link[=setup_x_axes]{setup_x_axes()}}, \code{\link[=setup_xy_axes]{setup_xy_axes()}}: Returns a grid::gTree
object containing the full axes.

\code{\link[=inject_inner_canvas_size]{inject_inner_canvas_size()}}: Returns the argument list with injected inner canvas
sizes.
}
\description{
\code{\link[=get_value_axes_width]{get_value_axes_width()}}: Get the width of the primary or secondary value
axes. The highest value will be put into its final format first, after that
the width is measured.

\code{\link[=get_variable_axes_height]{get_variable_axes_height()}}: Get the height of the variable axes.

\code{\link[=get_values_width]{get_values_width()}}: Get the width of the segment values.

\code{\link[=get_values_height]{get_values_height()}}: Get the height of the segment values.

\code{\link[=swap_xy_scaling]{swap_xy_scaling()}}: Width and height are measured according to the individual scaling
for each dimension. This function converts the measured to the respective other dimension.

\code{\link[=get_group_tick_positions_x]{get_group_tick_positions_x()}}: Calculates the x axes tick positions. In this
calculation ticks are meant to enclose the group labels. Meaning ticks are drawn
between groups.

\code{\link[=get_y_axes_values]{get_y_axes_values()}}: Generates a vector containing the values drawn on the
y axes. Values are equally spaced between the minimum and maximum value.

\code{\link[=setup_y_axes]{setup_y_axes()}}: Uses the currently active viewport to set up the y axes. The
primary axes is drawn on the left, the secondary axes on the right side of the
currently active viewport.

\code{\link[=setup_x_axes]{setup_x_axes()}}: Uses the currently active viewport to set up the x axes. The
height at which it is drawn is determined by the primary y axes 0 point. If y0
is not on the y axes then the x axes will either be drawn at the bottom, if only
positive values are present, or at the top, if only negative values are present.

\code{\link[=setup_xy_axes]{setup_xy_axes()}}: Uses the currently active viewport to set up both the x and
y axes.

\code{\link[=inject_inner_canvas_size]{inject_inner_canvas_size()}}: Injects the inner canvas size, meaning diagram size
minus axes size, into the dimensions parameter.
}
