% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{axes}
\alias{axes}
\alias{get_value_axes_width}
\alias{get_variable_axes_dimension}
\alias{get_group_tick_positions_x}
\alias{get_y_axes_values}
\alias{setup_y_axes}
\alias{setup_x_axes}
\alias{setup_xy_axes}
\title{Measure And Generate Custom Axes}
\usage{
get_value_axes_width(
  graphic_tab,
  max_value,
  axes,
  dimensions,
  visuals,
  which = "primary"
)

get_variable_axes_dimension(wrapped_text, dimensions, visuals)

get_group_tick_positions_x(
  number_of_groups,
  number_of_segments,
  segment_width,
  segment_pos
)

get_y_axes_values(values, axes, which = "primary")

setup_y_axes(tick_positions, tick_values, arguments, secondary = FALSE)

setup_x_axes(tick_positions, label_positions, labels, zero_pos, arguments)

setup_xy_axes(segment_info, arguments)
}
\arguments{
\item{graphic_tab}{The base data frame for the graphic.}

\item{max_value}{The maximum value which takes the most space.}

\item{axes}{The list of axes parameters.}

\item{dimensions}{The list of dimensions parameters.}

\item{visuals}{The list of visual parameters.}

\item{which}{Primary or secondary axes.}

\item{wrapped_text}{The wrapped text for the variable axes to measure.}

\item{number_of_groups}{The number of groups.}

\item{number_of_segments}{The number of segments per group.}

\item{segment_width}{The width of a single segment}

\item{segment_pos}{A vector containing all segment positions.}

\item{values}{Value vector from the data frame.}

\item{tick_positions}{The position of the axes ticks.}

\item{tick_values}{The values drawn at the tick positions.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{secondary}{FALSE by default. If TRUE draws the primary axes on the left side,
otherwise the secondary axes on the right side.}

\item{label_positions}{The positions the labels will be drawn at.}

\item{labels}{The labels per tick or per segment group.}

\item{zero_pos}{The y axes 0 position at which the x axes is drawn.}

\item{segment_info}{The list of measurements generated by
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}.}
}
\value{
\code{\link[=get_value_axes_width]{get_value_axes_width()}}: Returns a numeric width.

\code{\link[=get_variable_axes_dimension]{get_variable_axes_dimension()}}: Returns a numeric height.

\code{\link[=get_group_tick_positions_x]{get_group_tick_positions_x()}}: Returns a vector of numeric tick positions.

\code{\link[=get_y_axes_values]{get_y_axes_values()}}: Returns a numeric vector of equally spaced round values.

\code{\link[=setup_y_axes]{setup_y_axes()}}, \code{\link[=setup_x_axes]{setup_x_axes()}}, \code{\link[=setup_xy_axes]{setup_xy_axes()}}: Returns a grid::gTree
object containing the full axes.
}
\description{
\code{\link[=get_value_axes_width]{get_value_axes_width()}}: Get the width of the primary or secondary value
axes. The highest value will be put into its final format first, after that
the width is measured.

\code{\link[=get_variable_axes_dimension]{get_variable_axes_dimension()}}: Get the height of the variable axes.

\code{\link[=get_group_tick_positions_x]{get_group_tick_positions_x()}}: Calculates the x axes tick positions. In this
calculation ticks are meant to enclose the group labels. Meaning ticks are drawn
between groups.

\code{\link[=get_y_axes_values]{get_y_axes_values()}}: Generates a vector containing the values drawn on the
y axes. Values are equally spaced between the minimum and maximum value.

\code{\link[=setup_y_axes]{setup_y_axes()}}: Uses the currently active viewport to set up the y axes. The
primary axes is drawn on the left, the secondary axes on the right side of the
currently active viewport.

\code{\link[=setup_x_axes]{setup_x_axes()}}: Uses the currently active viewport to set up the x axes. The
height at which it is drawn is determined by the primary y axes 0 point. If y0
is not on the y axes then the x axes will either be drawn at the bottom, if only
positive values are present, or at the top, if only negative values are present.

\code{\link[=setup_xy_axes]{setup_xy_axes()}}: Uses the currently active viewport to set up both the x and
y axes.
}
