% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_es_cohens_d.R
\name{calc_es_d}
\alias{calc_es_d}
\title{Calculate Effect Size Cohen's d}
\usage{
calc_es_d(
  outcome,
  group,
  sd_pooled = NULL,
  es_type = "hedges",
  g_correction = "standard",
  conf_level = 0.95,
  paired = FALSE,
  bs_unique = NULL,
  within_vars = NULL
)
}
\arguments{
\item{outcome}{A vector of the outcome values.}

\item{group}{A vector of the grouping values.}

\item{sd_pooled}{Optional numeric value of the pooled standard deviation. If
not specified, the pooled standard deviation is calculated from group 1 and
2's standard deviations.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{g_correction}{Character indicating which Hedges' g correction to apply.
Options include \code{"standard"} for regular Hedges' g correction and
\code{"log"} to apply logarithmic Hedges' g correction. The default is
standard.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{paired}{A logical reflecting whether the pairwise comparison is between
(FALSE) or within (TRUE) subjects. If within, then the pair is tested
whether it is truly within by iterating through the unique between subjects
combinations. \code{bs_unique} must be specified.}

\item{bs_unique}{A vector of characters indicating the unique levels of the
between subjects variables.}

\item{within_vars}{A vector of characters indicating the repeated measures
variables.}
}
\description{
Calculate the Cohen's d, Hedges' g, or Glass's delta effect size
  for independent samples t-tests.
}
