% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plssem_prioritization.R
\name{calc_plssem_prioritization}
\alias{calc_plssem_prioritization}
\title{Partial Least Squares Structural Equation Modeling Prioritization}
\usage{
calc_plssem_prioritization(
  loadings = NULL,
  weights = NULL,
  data,
  total_estimates,
  outcomes,
  weight_type = "combo",
  cor_method = "pearson"
)
}
\arguments{
\item{loadings}{}

\item{weights}{}

\item{data}{}

\item{total_estimates}{}

\item{outcomes}{}

\item{weight_type}{}

\item{cor_method}{}
}
\description{
Partial Least Squares Structural Equation Modeling Prioritization
}
