% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_colors.R
\docType{data}
\name{color_themes}
\alias{color_themes}
\alias{add_color_theme}
\alias{get_theme_colors}
\alias{reset_color_themes}
\alias{display_colors}
\alias{display_themes}
\alias{override_theme}
\alias{.contrast_usage}
\title{Managing Global Color Themes}
\format{
An object of class \code{list} of length 10.
}
\usage{
add_color_theme(
  theme_name,
  base_colors,
  font_inside_colors = c("#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#000000",
    "#000000", "#000000", "#000000", "#000000", "#000000"),
  font_outside_colors = c("#000000", "#000000", "#000000", "#000000", "#000000",
    "#000000", "#000000", "#000000", "#000000", "#000000")
)

get_theme_colors(theme_name)

reset_color_themes(clear_themes = FALSE)

display_colors(theme_name)

display_themes()

override_theme(
  number = NULL,
  color = NULL,
  border_color = NULL,
  font_color = NULL
)

.contrast_usage
}
\arguments{
\item{theme_name}{The name of a globally stored theme.}

\item{base_colors}{The base colors for the segments.}

\item{font_inside_colors}{The individual font colors used when drawing values inside segments
on the corresponding base color.}

\item{font_outside_colors}{The individual font colors used when drawing values outsidr segments
on the corresponding base color.}

\item{clear_themes}{FALSE by default. If TRUE clears the global theme list.}

\item{number}{The number of the segment which should be manipulated.}

\item{color}{The new main segment color.}

\item{border_color}{The new border color of the segment.}

\item{font_color}{The new font color of the segment value.}
}
\value{
\code{\link[=add_color_theme]{add_color_theme()}}: Returns modified global theme list.

\code{\link[=get_theme_colors]{get_theme_colors()}}: A list with three vectors of hex color codes.

\code{\link[=reset_color_themes]{reset_color_themes()}}: Default global color list or empty list.

\code{\link[=display_colors]{display_colors()}}: Returns a list of base and font colors.

\code{\link[=display_themes]{display_themes()}}: A list containing all globally stored themes with their color
codew vectors.

\code{\link[=override_theme]{override_theme()}}: A list containing the override parameters.
}
\description{
\code{\link[=add_color_theme]{add_color_theme()}}: Adds a new color theme containing base and font colors to the
global theme list.

\code{\link[=get_theme_colors]{get_theme_colors()}}: Retrieve a list with three vectors of hex colors from the
globally set up themes containing the base colors for the segments and the
corresponding font colors.

\code{\link[=reset_color_themes]{reset_color_themes()}}: Resets the global color themes back to default.

\code{\link[=display_colors]{display_colors()}}: Displays all colors of a theme with the corresponding hex codes.

\code{\link[=display_themes]{display_themes()}}: Display all colors of all themes currently present in the
global environment.

\code{\link[=override_theme]{override_theme()}}: Is used to override certain elements of an individual segment.
Can be used to e.g. give certain segments a special coloring.
}
\examples{
# Adding and getting color themes
add_color_theme(theme_name  = "tropic",
                base_colors = c("#1B3A2E", "#1F5A3F", "#257F54", "#2FA36A", "#56C07E",
                                "#85D49A", "#B2E4B8", "#D7F0D5", "#ECF8EB", "#F7FCF7"),
                font_inside_colors = c("#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#000000",
                                        "#000000", "#000000", "#000000", "#000000", "#000000"),
                font_outside_colors = c("#1B3A2E", "#1F5A3F", "#257F54", "#2FA36A", "#56C07E",
                                        "#85D49A", "#B2E4B8", "#D7F0D5", "#ECF8EB", "#F7FCF7"))

get_theme_colors("ocean")

base_colors  <- get_theme_colors("ocean")[[1]]
font_inside  <- get_theme_colors("ocean")[[2]]
font_outside <- get_theme_colors("ocean")[[3]]

# Clear all globally stored themes in case only self created themes should be stored
reset_color_themes(clear_themes = TRUE)

# Reset global themes to default
reset_color_themes()

# Displaying colors and themes
display_colors("ocean")

display_themes()

# Example data frame
my_data <- dummy_data(100)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 25" = 18:24,
    "25 to under 55" = 25:54,
    "55 to under 65" = 55:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Male"   = 1,
    "Female" = 2)

# Override specific segment visuals to make them stand out
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "age",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.),
                    visuals        = graphic_visuals(
                        color_theme    = "violet_fire",
                        theme_override = list(override_theme(4, "#FF0000", "#00FF00", "#000000"),
                                              override_theme(7, "#00FFFF", "#FFFF00", "#0000FF"))))

# The two main color usages are set up like this
contrast_usage <- list(c(1),
                       c(1, 3),
                       c(1, 3, 5),
                       c(1, 3, 5, 7),
                       c(1, 3, 5, 7, 9),
                       c(1, 2, 3, 5, 7, 9),
                       c(1, 2, 3, 4, 5, 7, 9),
                       c(1, 2, 3, 4, 5, 6, 7, 9),
                       c(1, 2, 3, 4, 5, 6, 7, 8, 9),
                       c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

sequential_usage <- list(c(1),
                         c(1, 2),
                         c(1, 2, 3),
                         c(1, 2, 3, 4),
                         c(1, 2, 3, 4, 5),
                         c(1, 2, 3, 4, 5, 6),
                         c(1, 2, 3, 4, 5, 6, 7),
                         c(1, 2, 3, 4, 5, 6, 7, 8),
                         c(1, 2, 3, 4, 5, 6, 7, 8, 9),
                         c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))

}
\seealso{
Main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=modify_graphic_visuals]{modify_graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}},
\code{\link[=modify_graphic_axes]{modify_graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}}, \code{\link[=modify_graphic_dimensions]{modify_graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=modify_graphic_output]{modify_graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_fine_tuning]{modify_graphic_fine_tuning()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}
}
\keyword{datasets}
