% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_helpers.R
\name{create_table_of_contents}
\alias{create_table_of_contents}
\title{Create Table Of Contents For 'Excel' Workbooks}
\usage{
create_table_of_contents(
  wb,
  style = excel_output_style(),
  toc_header = "Table of Contents",
  toc_sheet_name = "Contents",
  titles = c(),
  subheaders = list(),
  subheader_colors = c(),
  subheader_underline = FALSE,
  colored_tabs = FALSE,
  file = NULL,
  print = .qol_options[["print"]]
)
}
\arguments{
\item{wb}{Workbook which receives the table of contents sheet.}

\item{style}{A list of options can be passed to control the appearance of the
table of contents. Styles can be created with \code{\link[=excel_output_style]{excel_output_style()}}.}

\item{toc_header}{The main header.}

\item{toc_sheet_name}{The table of contents sheet name.}

\item{titles}{Input a vector of custom titles displayed along the sheet names.}

\item{subheaders}{A list of custom subheaders. The entry names are the actual
subheaders to be displayed, while the values are the sheet names at which the
subheaders start.}

\item{subheader_colors}{Subheader background colors. These colors will also be used
to color the tabs, if the option is activated.}

\item{subheader_underline}{FALSE by default. If TRUE underlines the subheaders.}

\item{colored_tabs}{FALSE by default. If TRUE colors the tabs according to the
subheader colors.}

\item{file}{If NULL, opens the output as temporary file. If a filename with path
is specified, saves the output to the specified path.}

\item{print}{TRUE by default. If TRUE prints the output, if FALSE doesn't print
anything. Can be used if one only wants to catch the combined workbook.}
}
\value{
Returns a modified 'Excel' workbook with table of contents sheet.
}
\description{
Creates a table of contents sheet with custom style based on the provided 'Excel'
workbook.
}
\examples{
# Example data frame
my_data <- dummy_data(10)

# First we create a workbook to play around with
# print = FALSE and output = "excel_nostyle".
# This skips the styling and output part, so that the function runs faster.
set_print(FALSE)
set_output("excel_nostyle")

set_style_options(sheet_name = "sheet1")
tab1 <- my_data |> export_with_style()

set_style_options(sheet_name = "sheet2")
tab2 <- my_data |> export_with_style()

set_style_options(sheet_name = "sheet3")
tab3 <- my_data |> export_with_style()

set_style_options(sheet_name = "sheet4")
tab4 <- my_data |> export_with_style()

set_style_options(sheet_name = "sheet5")
tab5 <- my_data |> export_with_style()

set_style_options(sheet_name = "sheet6")
tab6 <- my_data |> export_with_style()

wb <- combine_into_workbook(tab1, tab2, tab3, tab4, tab5, tab6)

# Now add a custom styled table of contents sheet
create_table_of_contents(wb,
                         toc_header     = "My Custom Header",
                         toc_sheet_name = "TOC",
                         titles         = c("A title for sheet number 1",
                                            "A second title, this one goes to number two",
                                            "And a third one",
                                            "By the way: If you don't input titles here",
                                            "The title column will be empty. See below."),
                         subheaders      = list("This is a subheadline" = "sheet1",
                                                "Another one"           = "sheet3",
                                                "And another one"       = "sheet5"),
                         subheader_underline = TRUE,
                         colored_tabs        = TRUE,
                         style               = excel_output_style(toc_header_font_size = 20,
                                                                  toc_other_font_size = 14))

# Reset the global options afterwards
set_print(TRUE)
set_output("excel")

}
\seealso{
Creating a custom table style: \code{\link[=excel_output_style]{excel_output_style()}}, \code{\link[=modify_output_style]{modify_output_style()}},
\code{\link[=number_format_style]{number_format_style()}}, \code{\link[=modify_number_formats]{modify_number_formats()}}.

Global style options: \code{\link[=set_style_options]{set_style_options()}}, \code{\link[=set_labels]{set_labels()}}.

Other global options: \code{\link[=set_titles]{set_titles()}}, \code{\link[=set_footnotes]{set_footnotes()}}, \code{\link[=set_print]{set_print()}}, \code{\link[=set_monitor]{set_monitor()}},
\code{\link[=set_na.rm]{set_na.rm()}}, \code{\link[=set_print]{set_print()}}, \code{\link[=set_print_miss]{set_print_miss()}}, \code{\link[=set_output]{set_output()}}.

Combine Excel workbooks: \code{\link[=combine_into_workbook]{combine_into_workbook()}}.

Creating formats: \code{\link[=discrete_format]{discrete_format()}} and \code{\link[=interval_format]{interval_format()}}.

Functions that can handle formats and styles: \code{\link[=frequencies]{frequencies()}}, \code{\link[=crosstabs]{crosstabs()}}.

Additional functions that can handle styles: \code{\link[=export_with_style]{export_with_style()}}

Additional functions that can handle formats: \code{\link[=summarise_plus]{summarise_plus()}}, \code{\link[=recode.]{recode.()}},
\code{\link[=recode_multi]{recode_multi()}}, \code{\link[=transpose_plus]{transpose_plus()}}, \code{\link[=sort_plus]{sort_plus()}}
}
