% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstabs.R
\name{crosstabs}
\alias{crosstabs}
\title{Display Cross Table of Two Variables}
\usage{
crosstabs(
  data_frame,
  rows,
  columns,
  show_total = TRUE,
  statistics = "sum",
  formats = c(),
  by = c(),
  weight = NULL,
  full_precision = FALSE,
  titles = .qol_options[["titles"]],
  footnotes = .qol_options[["footnotes"]],
  style = .qol_options[["excel_style"]],
  output = .qol_options[["output"]],
  na.rm = .qol_options[["na.rm"]],
  print_miss = .qol_options[["print_miss"]],
  print = .qol_options[["print"]],
  monitor = .qol_options[["monitor"]]
)
}
\arguments{
\item{data_frame}{A data frame in which are the variables to tabulate.}

\item{rows}{The variable that appears in the table rows.}

\item{columns}{The variable that appears in the table columns.}

\item{show_total}{TRUE by default. Whether to print row and column totals or not.}

\item{statistics}{The user requested statistics.Available functions:
\itemize{
\item "sum"        -> Weighted and unweighted sum
\item "freq"       -> Unweighted frequency
\item "pct_row"    -> Weighted and unweighted row percentages
\item "pct_column" -> Weighted and unweighted column percentages
\item "pct_total"  -> Weighted and unweighted percentages compared to the grand total
}}

\item{formats}{A list in which is specified which formats should be applied to which variables.}

\item{by}{Compute tables stratified by the expressions of the provided variables.}

\item{weight}{Put in a weight variable to compute weighted results.}

\item{full_precision}{FALSE by default. If TRUE, the rounding of the values according
to the number formats in the style parameter is skipped and all values are output with all
their decimal places.}

\item{titles}{Specify one or more table titles. If you want to add hyperlinks you
can do so by adding "link:" followed by the hyperlink to the main text. To link to a file
use "file:" and pass the full file path afterwards. Linking to another cell works with "cell:".
When linking to a cell you can use to following keywords to automatically link to specific
parts of the output: first_title, last_title, first_footnote, last_footnote, table_start,
table_end, table_header.}

\item{footnotes}{Specify one or more table footnotes. If you want to add hyperlinks you
can do so by adding "link:" followed by the hyperlink to the main text. To link to a file
use "file:" and pass the full file path afterwards. Linking to another cell works with "cell:".
When linking to a cell you can use to following keywords to automatically link to specific
parts of the output: first_title, last_title, first_footnote, last_footnote, table_start,
table_end, table_header.}

\item{style}{A list of options can be passed to control the appearance of 'Excel' outputs.
Styles can be created with \code{\link[=excel_output_style]{excel_output_style()}}.}

\item{output}{The following output formats are available: console (default), text,
excel and excel_nostyle.}

\item{na.rm}{FALSE by default. If TRUE removes all NA values from the variables.}

\item{print_miss}{FALSE by default. If TRUE outputs all possible categories of the
grouping variables based on the provided formats, even if there are no observations
for a combination.}

\item{print}{TRUE by default. If TRUE prints the output, if FALSE doesn't print anything. Can be used
if one only wants to catch the output data frame.}

\item{monitor}{FALSE by default. If TRUE, outputs two charts to visualize the functions time consumption.}
}
\value{
Returns a data tables containing the results for the cross table.
}
\description{
\code{\link[=crosstabs]{crosstabs()}} produces a cross table of two variables. Statistics can be
weighted sums, unweighted frequencies or different percentages.
}
\details{
\code{\link[=crosstabs]{crosstabs()}} is based on the 'SAS' procedure Proc Freq, which provides
efficient and readable ways to perform cross tabulations.

To create a cross table you only need to provide a variable for the rows and columns.
Nothing special about this. The real power comes into play, when you output your
tables as a fully styled 'Excel' workbook. Setting up a custom, reusable style is
as easy as setting up options like: provide a color for the table header, set the
font size for the row header, should borders be drawn for the table cells yes/no,
and so on.

You can not only output sums and frequencies, but also different percentages, all
set up in separate, evenly designed tables. For just a quick overview, rather than
fully designed tables, you can also just output the tables in ASCII style format.
}
\examples{
# NOTE: The global option for printing the output is set to FALSE to save time
#       on running the examples. When running the examples manually, don't run
#       this option here.
#       Background color is also removed to make the code run faster.
set_print(FALSE)
set_style_options(background_color = "")

# Example data frame
my_data <- dummy_data(1000)

# Define titles and footnotes.
set_titles("This is title number 1",
           "This is title number 2",
           "This is title number 3")

set_footnotes("This is footnote number 1",
              "This is footnote number 2",
              "This is footnote number 3")

# Output cross tables
my_data |> crosstabs(state, sex)
my_data |> crosstabs(state, sex,
                     weight = "weight")

# Also works with characters
my_data |> crosstabs("state", "sex")
my_data |> crosstabs("state", "sex",
                     weight = "weight")

# Applying formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 25" = 18:24,
    "25 to under 55" = 25:54,
    "55 to under 65" = 55:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Total"  = 1:2,
    "Male"   = 1,
    "Female" = 2)

my_data |> crosstabs(age, sex,
                     formats = list(age = age., sex = sex.))

# Split cross table by expressions of another variable
my_data |> crosstabs(state, sex, by = education)

# Compute different stats
my_data |> crosstabs(state, sex,
                     statistics = c("sum", "freq", "pct_row", "pct_column", "pct_total"))

# Get a list with two data tables for further usage
result_list <- my_data |> crosstabs(age, sex,
                                    formats = list(age = age., sex = sex.))

# Output in text file
my_data |> crosstabs(state, sex, output = "text")

# Output to Excel
my_data |> crosstabs(state, sex, output = "excel")

# If you want to add hyperlinksto titles and footnotes you can do so by
# adding "link:" followed by the hyperlink to the main text. Linking to another
# cell works with "cell:". To link to a file use "file:" and pass the full file
# path afterwards. When linking to a cell you can use to following keywords to
# automatically link to specific parts of the output: first_title, last_title,
# first_footnote, last_footnote, table_start, table_end, table_header.
set_titles("This is title number 1",
           "This is title number 2 link: https://cran.r-project.org/",
           "This is title number 3 cell: W22",
           "This is title number 4 file: C:/MyFolder/MyFile.txt")

set_footnotes("This is footnote number 1",
              "This is footnote number 2 file: C:/MyFolder/MyFile.txt",
              "This is footnote number 3 cell: W22",
              "This is footnote number 4 link: https://cran.r-project.org/")

# Individual styling can also be passed directly
my_style <- excel_output_style(header_back_color = "0077B6",
                               font              = "Times New Roman")

my_data |> crosstabs(age, sex, output = "excel", style = my_style,
                     formats = list(age = age., sex = sex.))

# To save a table as xlsx file you have to set the path and filename in the
# style element
# Example files paths
table_file <- tempfile(fileext = ".xlsx")

# Note: Normally you would directly input the path ("C:/MyPath/") and name ("MyFile.xlsx").
#       With the set_style_options you can also set a table style globally.
set_style_options(save_path  = dirname(table_file),
                  file       = basename(table_file),
                  sheet_name = "MyTable")

my_data |> crosstabs(state, sex, output = "excel")

# Manual cleanup for example
unlink(table_file)

# Global options are permanently active until the current R session is closed.
# There are also functions to reset the values manually.
reset_qol_options()

}
\seealso{
Creating a custom table style: \code{\link[=excel_output_style]{excel_output_style()}}, \code{\link[=modify_output_style]{modify_output_style()}},
\code{\link[=number_format_style]{number_format_style()}}, \code{\link[=modify_number_formats]{modify_number_formats()}}.

Global style options: \code{\link[=set_style_options]{set_style_options()}}, \code{\link[=set_labels]{set_labels()}}.

Other global options: \code{\link[=set_titles]{set_titles()}}, \code{\link[=set_footnotes]{set_footnotes()}}, \code{\link[=set_print]{set_print()}}, \code{\link[=set_monitor]{set_monitor()}},
\code{\link[=set_na.rm]{set_na.rm()}}, \code{\link[=set_print]{set_print()}}, \code{\link[=set_print_miss]{set_print_miss()}}, \code{\link[=set_output]{set_output()}}.

Combine Excel workbooks: \code{\link[=combine_into_workbook]{combine_into_workbook()}}.

Standalone table of contents: \code{\link[=create_table_of_contents]{create_table_of_contents()}}.

Creating formats: \code{\link[=discrete_format]{discrete_format()}} and \code{\link[=interval_format]{interval_format()}}.

Functions that can handle formats and styles: \code{\link[=frequencies]{frequencies()}}, \code{\link[=any_table]{any_table()}}.

Additional functions that can handle styles: \code{\link[=export_with_style]{export_with_style()}}

Additional functions that can handle formats: \code{\link[=summarise_plus]{summarise_plus()}}, \code{\link[=recode.]{recode.()}},
\code{\link[=recode_multi]{recode_multi()}}, \code{\link[=transpose_plus]{transpose_plus()}}, \code{\link[=sort_plus]{sort_plus()}}
}
