% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaphic_design.R
\name{design_graphic}
\alias{design_graphic}
\title{Generate Graphics While Fully Concentrating On Design}
\usage{
design_graphic(
  data_frame,
  axes_variables,
  segments,
  values,
  statistics = "pct_group",
  pct_value = list(),
  formats = list(),
  by = c(),
  weight = NULL,
  titles = .qol_options[["titles"]],
  footnotes = .qol_options[["footnotes"]],
  var_labels = .qol_options[["var_labels"]],
  stat_labels = .qol_options[["stat_labels"]],
  diagram = vbars,
  colors = "ocean",
  color_usage = .qol_options[["graphic_color_usage"]],
  visuals = .qol_options[["graphic_visuals"]],
  axes = .qol_options[["graphic_axes"]],
  dimensions = .qol_options[["graphic_dimensions"]],
  number_formats = .qol_options[["graphic_number_formats"]],
  add_texts = .qol_options[["graphic_texts"]],
  output = .qol_options[["graphic_output"]],
  na.rm = .qol_options[["na.rm"]],
  print_miss = .qol_options[["print_miss"]],
  print = .qol_options[["print"]],
  monitor = .qol_options[["monitor"]]
)
}
\arguments{
\item{data_frame}{A data frame in which are the variables to tabulate.}

\item{axes_variables}{A vector that provides single variables or variable combinations
that should appear on the variable axes. To nest variables use the form:
"var1 + var2 + var3 + ...".}

\item{segments}{A vector that provides variables that should appear as the graphics
segments.}

\item{values}{A vector containing all variables that should be summarised.}

\item{statistics}{Available functions:
\itemize{
\item "sum"       -> Weighted and unweighted sum
\item "sum_wgt"   -> Sum of all weights
\item "freq"      -> Unweighted frequency
\item "freq_g0"   -> Unweighted frequency of all values greater than zero
\item "pct_group" -> Weighted and unweighted percentages within the respective group
\item "pct_value" -> Weighted and unweighted percentages between value variables
\item "pct_total" -> Weighted and unweighted percentages compared to the grand total
\item "mean"      -> Weighted and unweighted mean
\item "median"    -> Weighted and unweighted median
\item "mode"      -> Weighted and unweighted mode
\item "min"       -> Minimum
\item "max"       -> Maximum
\item "sd"        -> Weighted and unweighted standard deviation
\item "variance"  -> Weighted and unweighted standard variance
\item "first"     -> First value
\item "last"      -> Last value
\item "pn"        -> Weighted and unweighted percentiles (any p1, p2, p3, ... possible)
\item "missing"   -> Missings generated by the value variables
}}

\item{pct_value}{If pct_value is specified in the statistics, you can pass a list here
which contains the information for a new variable name and between which of the value
variables percentages should be computed.}

\item{formats}{A list in which is specified which formats should be applied to which variables.}

\item{by}{Compute tables stratified by the expressions of the provided variables.}

\item{weight}{Put in a weight variable to compute weighted results.}

\item{titles}{Specify one or more table titles.}

\item{footnotes}{Specify one or more table footnotes.}

\item{var_labels}{A list in which is specified which label should be printed for
which variable instead of the variable name.}

\item{stat_labels}{A list in which is specified which label should be printed for
which statistic instead of the statistic name.}

\item{diagram}{Function that provides the generation of the main diagram area.}

\item{colors}{The color theme to use. Can either be the name of a built-in or
custom theme or a vector of hex codes. Existing color themes can be viewed with
\code{\link[=display_themes]{display_themes()}}.}

\item{color_usage}{A list of numerical vectors that specify which specific colors
from the color scheme are used for which number of segments.}

\item{visuals}{Visual parameters set with \code{\link[=graphic_visuals]{graphic_visuals()}}.}

\item{axes}{Axes parameters set with \code{\link[=graphic_axes]{graphic_axes()}}.}

\item{dimensions}{Dimension parameters set with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.}

\item{number_formats}{Number formats set with \code{\link[=number_format_style]{number_format_style()}}.}

\item{add_texts}{Use the \code{\link[=add_textbox]{add_textbox()}} function to freely place one or multiple textboxes.}

\item{output}{Output parameters set with \code{\link[=graphic_output]{graphic_output()}}.}

\item{na.rm}{FALSE by default. If TRUE removes all NA values from the variables.}

\item{print_miss}{FALSE by default. If TRUE outputs all possible categories of the
grouping variables based on the provided formats, even if there are no observations
for a combination.}

\item{print}{TRUE by default. If TRUE prints the output, if FALSE doesn't print anything. Can be used
if one only wants to catch the output data frame and workbook with meta information.}

\item{monitor}{FALSE by default. If TRUE, outputs two charts to visualize the functions time consumption.}

\item{pct_group}{If pct_group is specified in the statistics, this option is used to
determine which variable of the row and column variables should add up to 100 \%.
Multiple variables can be specified in a vector to generate multiple group percentages.}
}
\value{
Returns a list with the data table containing the results for the graphic and the
final grid::gTree graphics object.
}
\description{
\code{\link[=design_graphic]{design_graphic()}} produces fully styled diagrams. The many parameters are all
focused around making visual changes simple and fast. All diagrams can be exported
into multiple formats.

The diagram parameter takes in a function for the main diagram area, making the
function modular and extendable.
}
\details{
\code{\link[=design_graphic]{design_graphic()}} is based on the 'SAS' procedure Proc Tabulate, which provides
efficient and readable ways to perform complex Generating graphics.
}
\examples{
# Example data frame
my_data <- dummy_data(1000)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 25" = 18:24,
    "25 to under 55" = 25:54,
    "55 to under 65" = 55:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Total"  = 1:2,
    "Male"   = 1,
    "Female" = 2)

# Design grouped vertical bar chart
qol_graphic <- my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex",
                    values         = weight,
                    diagram        = vbars,
                    formats        = list(sex = sex., age = age.),
                    titles         = "The qol-package brings powerful concepts
                                      from 'SAS' to 'R' to make life easier and
                                      produce bigger and more complex outputs in
                                      less time with less code.",
                    footnotes      = "The qol-package brings powerful concepts
                                      from 'SAS' to 'R' to make life easier and
                                      produce bigger and more complex outputs in
                                      less time with less code.")


}
\seealso{
Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=modify_graphic_visuals]{modify_graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}},
\code{\link[=modify_graphic_axes]{modify_graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}}, \code{\link[=modify_graphic_dimensions]{modify_graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=modify_graphic_output]{modify_graphic_output()}}

Global graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=set_color_theme]{set_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}

Main diagrams: \code{\link[=vbars]{vbars()}}

View colors and themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}

Mid level functions:
\itemize{
\item Textboxes: \code{\link[=add_textbox]{add_textbox()}}, \code{\link[=add_title]{add_title()}}, \code{\link[=add_footnote]{add_footnote()}}, \code{\link[=add_graphic_origin]{add_graphic_origin()}},
\code{\link[=wrap_single_text]{wrap_single_text()}}, \code{\link[=wrap_text_vector]{wrap_text_vector()}}, \code{\link[=register_windows_font]{register_windows_font()}}
\item Viewport: \code{\link[=setup_main_canvas]{setup_main_canvas()}}, \code{\link[=setup_nested_viewport]{setup_nested_viewport()}}, \code{\link[=setup_diagram_viewport]{setup_diagram_viewport()}},
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}, \code{\link[=back_to_the_root]{back_to_the_root()}}, \code{\link[=is_viewport_pushed]{is_viewport_pushed()}}
\item Dimensions: \code{\link[=get_available_width]{get_available_width()}}, \code{\link[=get_available_height]{get_available_height()}}
\item Diagram: \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}, \code{\link[=vbar_grob]{vbar_grob()}}
\item Axes: \code{\link[=get_value_axes_width]{get_value_axes_width()}}, \code{\link[=get_variable_axes_dimension]{get_variable_axes_dimension()}}, \code{\link[=get_group_tick_positions_x]{get_group_tick_positions_x()}},
\code{\link[=get_y_axes_values]{get_y_axes_values()}}, \code{\link[=setup_y_axes]{setup_y_axes()}}, \code{\link[=setup_x_axes]{setup_x_axes()}}, \code{\link[=setup_xy_axes]{setup_xy_axes()}}
\item Labels: \code{\link[=direct_vertical_labels]{direct_vertical_labels()}}
\item Output: \code{\link[=output_graphic]{output_graphic()}}
\item Other: \code{\link[=format_values]{format_values()}}
}

Creating formats: \code{\link[=discrete_format]{discrete_format()}} and \code{\link[=interval_format]{interval_format()}}.

Functions that can handle formats and styles: \code{\link[=frequencies]{frequencies()}}, \code{\link[=crosstabs]{crosstabs()}}, \code{\link[=any_table]{any_table()}}.

Additional functions that can handle styles: \code{\link[=export_with_style]{export_with_style()}}

Additional functions that can handle formats: \code{\link[=summarise_plus]{summarise_plus()}}, \code{\link[=recode]{recode()}},
\code{\link[=recode_multi]{recode_multi()}}, \code{\link[=transpose_plus]{transpose_plus()}}, \code{\link[=sort_plus]{sort_plus()}}
}
