% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{diagram}
\alias{diagram}
\alias{get_diagram_dimensions}
\alias{vbar_grob}
\title{Get The Dimensions And Positions Of Groups And Segments}
\usage{
get_diagram_dimensions(
  graphic_tab,
  axes_vars,
  segment_vars,
  values,
  axes = .qol_options[["graphic_axes"]],
  dimensions = .qol_options[["graphic_dimensions"]],
  visuals = .qol_options[["graphic_visuals"]]
)

vbar_grob(diagram_info, arguments, theme)
}
\arguments{
\item{graphic_tab}{The base data frame for the graphic.}

\item{axes_vars}{The names of the axes variables.}

\item{segment_vars}{The names of the segment variables.}

\item{values}{Value vector from the data frame.}

\item{axes}{The list of axes parameters.}

\item{dimensions}{The list of dimensions parameters.}

\item{visuals}{The list of visual parameters.}

\item{diagram_info}{The list of measurements generated by
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{theme}{Color theme used for the segments.}
}
\value{
Returns a list of parameters for groups, segments and axes.

Returns a grid::rectGrob object.
}
\description{
\code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}: Get various parameters concerning the division and
position of groups and segments. These are e.g. number of groups/segments, width of
groups/segments, ids or positions among other things.

\code{\link[=vbar_grob]{vbar_grob()}}: Set up the main segments for vertical bars.
}
