% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{dummy_data}
\alias{dummy_data}
\title{Dummy Data}
\usage{
dummy_data(no_obs, monitor = .qol_options[["monitor"]])
}
\arguments{
\item{no_obs}{Number of observations.}

\item{monitor}{FALSE by default. If TRUE outputs two charts to visualize the functions time consumption.}
}
\value{
Returns a dummy data table.
}
\description{
The dummy data frame contains a few randomly generated variables like year, sex,
age, income and weight to test out functionalities. It can be generated with the desired
number of observations.
}
\examples{
my_data <- dummy_data(1000)

}
