% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{format_values}
\alias{format_values}
\title{Format Values}
\usage{
format_values(
  values,
  decimals = .qol_options[["graphic_axes"]][["primary_axes_decimals"]],
  big_mark = .qol_options[["graphic_axes"]][["primary_axes_big_mark"]],
  decimal_mark = .qol_options[["graphic_axes"]][["primary_axes_decimal_mark"]],
  prefix = .qol_options[["graphic_axes"]][["primary_axes_prefix"]],
  suffix = .qol_options[["graphic_axes"]][["primary_axes_suffix"]],
  scale = .qol_options[["graphic_axes"]][["primary_axes_scale"]]
)
}
\arguments{
\item{values}{Single value or vector of values.}

\item{decimals}{Number of decimals to be displayed.}

\item{big_mark}{Character used as the thousand separator.}

\item{decimal_mark}{Character used as the decimal separator.}

\item{prefix}{What to put in front of the value.}

\item{suffix}{What to put after the value.}

\item{scale}{Multiplier for values.}
}
\value{
Returns a formatted value as character.
}
\description{
Format values to output numbers in a certain way.
}
