% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_axes}
\alias{graphic_axes}
\alias{modify_graphic_axes}
\title{Graphic Axes}
\usage{
graphic_axes(
  primary_axes_max = "auto",
  primary_axes_min = "auto",
  primary_axes_steps = 5,
  primary_axes_decimals = 0,
  primary_axes_big_mark = ".",
  primary_axes_decimal_mark = ",",
  primary_axes_prefix = "",
  primary_axes_suffix = "",
  primary_axes_scale = 1,
  primary_values_decimals = 1,
  primary_values_big_mark = ".",
  primary_values_decimal_mark = ",",
  primary_values_prefix = "",
  primary_values_suffix = "",
  secondary_axes_max = "auto",
  secondary_axes_min = "auto",
  secondary_axes_steps = 5,
  secondary_axes_decimals = 0,
  secondary_axes_big_mark = ".",
  secondary_axes_decimal_mark = ",",
  secondary_axes_prefix = "",
  secondary_axes_suffix = "",
  secondary_axes_scale = 1,
  secondary_values_decimals = 0,
  secondary_values_big_mark = ".",
  secondary_values_decimal_mark = ",",
  secondary_values_prefix = "",
  secondary_values_suffix = "",
  variable_axes_interval = 1
)

modify_graphic_axes(axes_to_modify, ...)
}
\arguments{
\item{primary_axes_max}{Maximum value for the primary axes. If "auto", the maximum
value is determined by the maximum value present in the graphic.}

\item{primary_axes_min}{Minimum value for the primary axes. If "auto", the minimum
value is determined by the minimum value present in the graphic. If there is no
negative value it will always be set to 0.}

\item{primary_axes_steps}{The number of steps from minimum to maximum value for
the primary axes.}

\item{primary_axes_decimals}{Number of decimal points the values have on the primary axes.}

\item{primary_axes_big_mark}{Character used as the thousand separator on the primary axes.}

\item{primary_axes_decimal_mark}{Character used as the decimal separator on the primary axes.}

\item{primary_axes_prefix}{What to put in front of the value on the primary axes.}

\item{primary_axes_suffix}{What to put after the value on the primary axes.}

\item{primary_axes_scale}{Multiplier for the values on the primary axes.}

\item{primary_values_decimals}{Number of decimal points the values plotted on the primary axes have.}

\item{primary_values_big_mark}{Character used as the thousand separator for the values plotted
on the primary axes have.}

\item{primary_values_decimal_mark}{Character used as the decimal separator for the values plotted
on the primary axes have.}

\item{primary_values_prefix}{What to put in front of the values plotted on the primary axes have.}

\item{primary_values_suffix}{What to put after the values plotted on the primary axes have.}

\item{secondary_axes_max}{Maximum value for the secondary axes. If "auto", the maximum
value is determined by the maximum value present in the graphic.}

\item{secondary_axes_min}{Minimum value for the secondary axes. If "auto", the minimum
value is determined by the minimum value present in the graphic. If there is no
negative value it will always be set to 0.}

\item{secondary_axes_steps}{The number of steps from minimum to maximum value for
the secondary axes.}

\item{secondary_axes_decimals}{Number of decimal points the values have on the secondary axes.}

\item{secondary_axes_big_mark}{Character used as the thousand separator on the secondary axes.}

\item{secondary_axes_decimal_mark}{Character used as the decimal separator on the secondary axes.}

\item{secondary_axes_prefix}{What to put in front of the value on the secondary axes.}

\item{secondary_axes_suffix}{What to put after the value on the secondary axes.}

\item{secondary_axes_scale}{Multiplier for values on the secondary axes.}

\item{secondary_values_decimals}{Number of decimal points the values plotted on the secondary axes have.}

\item{secondary_values_big_mark}{Character used as the thousand separator for the values plotted
on the secondary axes.}

\item{secondary_values_decimal_mark}{Character used as the decimal separator for the values plotted
on the secondary axes.}

\item{secondary_values_prefix}{What to put in front of the values plotted on the secondary axes.}

\item{secondary_values_suffix}{What to put after the values plotted on the secondary axes.}

\item{variable_axes_interval}{The number of steps in which to display labels on the variable axes.}

\item{axes_to_modify}{A pre created graphic axes object where only
certain elements should be modified while the rest is kept as is.}

\item{...}{Pass in names and corresponding new values for existing graphic elements.}
}
\value{
Returns a list of named graphic options.

Returns a modified list of named graphic options.
}
\description{
Set different options regarding the axes of a graphic produced by
\code{\link[=design_graphic]{design_graphic()}}. All parameters can also be set globally via \code{\link[=set_graphic_options]{set_graphic_options()}}.

If options are stored in an object, it can be modified afterwards. In this case
only the options to modify will be adjusted while the rest stays as is.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# For default values
default_axes <- graphic_axes()

# Set specific options, the rest will be set to default values
custom_axes <- graphic_axes(primary_axes_max      = 100,
                            primary_axes_decimals = 1)

# Apply options locally
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vbars,
                    axes           = custom_axes)

# Or direct
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vbars,
                    axes           = graphic_axes(primary_axes_max      = 100,
                                                  primary_axes_decimals = 1))

# Set options globally
set_graphic_options(primary_axes_max      = 100,
                    primary_axes_decimals = 1)

my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vbars)

# Modify the previously created graphic options
custom_axes <- custom_axes |> modify_graphic_axes(primary_axes_min = 50)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=modify_graphic_visuals]{modify_graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}},
\code{\link[=modify_graphic_axes]{modify_graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}}, \code{\link[=modify_graphic_dimensions]{modify_graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=modify_graphic_output]{modify_graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_fine_tuning]{modify_graphic_fine_tuning()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}

Color themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}, \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}, \code{\link[=override_theme]{override_theme()}}
}
