% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_dimensions}
\alias{graphic_dimensions}
\title{Graphic Dimensions}
\usage{
graphic_dimensions(
  graphic_width = 16,
  graphic_height = 9,
  diagram_start = "auto",
  diagram_width = "auto",
  diagram_height = "auto",
  margins = 0.25,
  title_font_size = 9,
  footnote_font_size = 8,
  axes_font_size = 9,
  value_font_size = 9,
  label_font_size = 9,
  origin_font_size = 8,
  other_font_size = 9,
  space_between_bars_pct = 0,
  bar_overlap = 0,
  line_thickness = 1,
  segment_line_length = 1,
  segment_line_offset = 0,
  textbox_width = 2
)
}
\arguments{
\item{graphic_width}{The width of the whole graphic.}

\item{graphic_height}{The height of the whole graphic.}

\item{diagram_start}{The starting position of the main diagram within the graphic from the top.}

\item{diagram_width}{The width of the main diagram within the graphic.}

\item{diagram_height}{The height of the main diagram within the graphic.}

\item{margins}{Inner margins to the graphic borders in cm.}

\item{title_font_size}{Font size of the title}

\item{footnote_font_size}{Font size of the footnote}

\item{axes_font_size}{Font size of the axes}

\item{value_font_size}{Font size of the values.}

\item{label_font_size}{Font size of the origin text.}

\item{origin_font_size}{Font size of the origin text.}

\item{other_font_size}{Font size of every other text element.}

\item{space_between_bars_pct}{The space between adjacent bars.}

\item{bar_overlap}{The factor determines how much bars are overlapping each other.}

\item{line_thickness}{The thickness of lines in points.}

\item{segment_line_length}{The length of the lines leading from segments to labels in cm.}

\item{segment_line_offset}{Offset in the height of the leading lines in cm. If 0, all leading lines
will end at the same height, meaning the labels will be drawn at the same height. If
an offset is set, all lines after the first one will be shorter by this amount * number
of the line. Enables the labels to be drawn in steps, if e.g. the labels have a long
text which would overlap, if they where on the same height.}

\item{textbox_width}{Determines the width of any freely placed textbox.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the dimensions of a graphic produced by \code{\link[=design_graphic]{design_graphic()}}.
}
\examples{
# For default values
gd <- graphic_dimensions()

# Set specific options, the rest will be set to default values
gd <- graphic_dimensions(graphic_width  = 10,
                         graphic_height = 10)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
