% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_fine_tuning}
\alias{graphic_fine_tuning}
\title{Graphic Fine Tuning}
\usage{
graphic_fine_tuning(
  diagram_margin = 0.01,
  values_vjust_positive = 1.7,
  values_vjust_negative = -0.7,
  values_vjust_90_positive = 1.2,
  values_vjust_90_negative = -0.2,
  value_overlap_factor = 1.35,
  shrink_segment_width = 0.55,
  values_hjust = 0.5,
  values_hjust_90 = 0.35,
  values_hjust_90_plus = 0.05,
  values_zero_line_offset = 0.2,
  values_below_axes_just = 1.5,
  values_below_axes_90_just = 1.4,
  tick_length = 0.02,
  value_axes_margin = 0.02,
  y_axes_scaling = 1.1,
  swap_direction_threshold = 0.75,
  segment_line_offset = 0.3,
  segment_line_treshhold = 0.95,
  segment_label_hjust = 0.05,
  max_segment_label_shift = 0.25,
  cm_to_inch_factor = 2.54,
  svg_anchor_adjust = 0.6,
  svg_line_height_adjust = 0.65,
  default_stack_width = 35,
  stacked_value_vjust = 0.4,
  segment_line_length_stacked = 0.05,
  stacked_totals_distance = 3
)
}
\arguments{
\item{diagram_margin}{Margin used within the diagram area.}

\item{values_vjust_positive}{Positive vertical adjustment for values in vbars.}

\item{values_vjust_negative}{Negative vertical adjustment for values in vbars.}

\item{values_vjust_90_positive}{Positive vertical adjustment for rotated values in vbars.}

\item{values_vjust_90_negative}{Negative vertical adjustment for rotated values in vbars.}

\item{value_overlap_factor}{Used as a multiplier for the value height. If the value height
multiplied by this factor exceeds the segment height of a vbar, the value will be automatically
drawn outside the segment.}

\item{shrink_segment_width}{Shrinks the segment width by this factor, if the segment borders
are colored, to prevent the bars from overlapping}

\item{values_hjust}{Horizontal adjustment of segment values.}

\item{values_hjust_90}{Horizontal adjustment of rotated segment values.}

\item{values_hjust_90_plus}{Additional horizontal adjustment of rotated segment values, if
drawn inside segments.}

\item{values_zero_line_offset}{Offset to the x axes for 0 values.}

\item{values_below_axes_just}{Adjustment for 0 values, if drawn below x axes.}

\item{values_below_axes_90_just}{Adjustment for rotated 0 values, if drawn below x axes.}

\item{tick_length}{Length of the axes ticks.}

\item{value_axes_margin}{Additional margin when measuring the value axes width}

\item{y_axes_scaling}{Scaling factor for the y axes, if maximum value is calculated
automatically.}

\item{swap_direction_threshold}{Determines the threshold for the variable axes at which
the drawing direction is swapped.}

\item{segment_line_offset}{A static offset at which segment lines are drawn.}

\item{segment_line_treshhold}{Segment lines won't be drawn further than the maximum
value multiplied by this factor.}

\item{segment_label_hjust}{Horizontal adjustment of the segment labels if they are drawn
in stairs.}

\item{max_segment_label_shift}{The maximum percentage of the textbox width the segment
labels can be shifted to achieve a decollision.}

\item{cm_to_inch_factor}{cm to inc conversion factor.,}

\item{svg_anchor_adjust}{Text anchoring adjustment for interactive SVG graphics, so that the
positions look like in the static version.}

\item{svg_line_height_adjust}{Text line height in interactive SVG graphics is bigger and needs
to be turned down.}

\item{default_stack_width}{The stack width is controlled via the space_between_bars parameter
inside \code{\link[=graphic_dimensions]{graphic_dimensions()}}. Since this value defaults to 0, it would completely collapse
the stacks. There this parameter jumps into place and overrides the default.}

\item{stacked_value_vjust}{Vertical adjustment of the values inside vertical stacked charts.}

\item{segment_line_length_stacked}{The length of the segment lines in stacked charts.}

\item{stacked_totals_distance}{Distance in mm of the total values from the stacks.}
}
\value{
Returns a list of named graphic options.
}
\description{
Throughout the graphic generation some fixed values are used. These can be adjusted
to make detailed graphical changes in certain areas. These values normally should
only be altered in edge cases. All parameters can also be set globally via \code{\link[=set_graphic_options]{set_graphic_options()}}.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# For default values
default_fine_tuning <- graphic_fine_tuning()

# Set specific options, the rest will be set to default values
custom_fine_tuning <- graphic_fine_tuning(values_vjust_positive = 2.5,
                                          tick_length           = 0.05)

# Apply options locally
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    fine_tuning    = custom_fine_tuning)

# Or direct
my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars,
                    fine_tuning    = graphic_fine_tuning(values_vjust_positive = 2.5,
                                                         tick_length           = 0.05))

# Set options globally
set_graphic_options(values_vjust_positive = 2.5,
                    tick_length           = 0.05)

my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "education",
                    diagram        = dg_vertical_bars)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Graphic options: \code{\link[=graphic_visuals]{graphic_visuals()}}, \code{\link[=graphic_axes]{graphic_axes()}}, \code{\link[=graphic_dimensions]{graphic_dimensions()}},
\code{\link[=graphic_output]{graphic_output()}}, \code{\link[=graphic_fine_tuning]{graphic_fine_tuning()}}, \code{\link[=modify_graphic_options]{modify_graphic_options()}}

Global graphic options: \code{\link[=set_graphic_options]{set_graphic_options()}}, \code{\link[=get_graphic_options]{get_graphic_options()}}, \code{\link[=reset_graphic_options]{reset_graphic_options()}}

Color themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}, \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_colors]{get_theme_colors()}},
\code{\link[=reset_color_themes]{reset_color_themes()}}, \code{\link[=override_theme]{override_theme()}}
}
