% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_fine_tuning}
\alias{graphic_fine_tuning}
\title{Graphic Fine Tuning}
\usage{
graphic_fine_tuning(
  line_height = 1.1,
  diagram_start_adjust = 2,
  diagram_height_adjust = 4,
  diagram_margin = 0.01,
  values_vjust_positive = 1.7,
  values_vjust_negative = -0.7,
  values_vjust_90_positive = 1.1,
  values_vjust_90_negative = -0.1,
  value_overlap_factor = 1.35,
  shrink_segment_width = 0.5,
  values_rotation = 90,
  values_hjust = 0.5,
  values_hjust_90 = 0.35,
  values_hjust_90_plus = 0.05,
  values_zero_line_offset = 0.2,
  values_below_axes_just = 1.5,
  values_below_axes_90_just = 1.4,
  tick_length = 0.02,
  value_axes_margin = 0.02,
  variable_axes_margin = 0.03,
  value_height_factor = 1.5,
  y_axes_scaling = 1.3,
  swap_direction_threshold = 0.75,
  segment_line_offset = 0.3,
  segment_line_treshhold = 0.95,
  segment_label_hjust = 0.05,
  max_segment_label_shift = 0.25,
  cm_to_inch_factor = 2.54
)
}
\arguments{
\item{line_height}{The height of a single text line.}

\item{diagram_start_adjust}{Adjusts the diagram starting position, when it is automatically
determined.}

\item{diagram_height_adjust}{Adjusts the diagram height, when it is automatically determined.}

\item{diagram_margin}{Margin used within the diagram area.}

\item{values_vjust_positive}{Positive vertical adjustment for values in vbars.}

\item{values_vjust_negative}{Negative vertical adjustment for values in vbars.}

\item{values_vjust_90_positive}{Positive vertical adjustment for rotated values in vbars.}

\item{values_vjust_90_negative}{Negative vertical adjustment for rotated values in vbars.}

\item{value_overlap_factor}{Used as a multiplier for the value height. If the value height
multiplied by this factor exceeds the segment height of a vbar, the value will be automatically
drawn outside the segment.}

\item{shrink_segment_width}{Shrinks the segment width by this factor, if the segment borders
are colored, to prevent the bars from overlapping}

\item{values_rotation}{Degrees of segment value rotation.}

\item{values_hjust}{Horizontal adjustment of segment values.}

\item{values_hjust_90}{Horizontal adjustment of rotated segment values.}

\item{values_hjust_90_plus}{Additional horizontal adjustment of rotated segment values, if
drawn inside segments.}

\item{values_zero_line_offset}{Offset to the x axes for 0 values.}

\item{values_below_axes_just}{Adjustment for 0 values, if drawn below x axes.}

\item{values_below_axes_90_just}{Adjustment for rotated 0 values, if drawn below x axes.}

\item{tick_length}{Length of the axes ticks.}

\item{value_axes_margin}{Additional margin when measuring the value axes width}

\item{variable_axes_margin}{Additional margin when measuring the variable axes height.}

\item{value_height_factor}{Scaling factor when measuring value widths.}

\item{y_axes_scaling}{Scaling factor for the y axes, if maximum value is calculated
automatically.}

\item{swap_direction_threshold}{Determines the threshold for the variable axes at which
the drawing direction is swapped.}

\item{segment_line_offset}{A static offset at which segment lines are drawn.}

\item{segment_line_treshhold}{Segment lines won't be drawn further than the maximum
value multiplied by this factor.}

\item{segment_label_hjust}{Horizontal adjustment of the segment labels if they are drawn
in stairs.}

\item{max_segment_label_shift}{The maximum percentage of the textbox width the segment
labels can be shifted to achieve a decollision.}

\item{cm_to_inch_factor}{cm to inc conversion factor.}
}
\value{
Returns a list of named graphic options.
}
\description{
Throughout the graphic generation some fixed values are used. These can be adjusted
to make detailed graphical changes in certain areas. These values normally should
only be altered in edge cases.
}
\examples{
# For default values
gft <- graphic_fine_tuning()

# Set specific options, the rest will be set to default values
gft <- graphic_fine_tuning(diagram_start_adjust = 2.5,
                           tick_length          = 0.05)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
