% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_output}
\alias{graphic_output}
\title{Graphic Output}
\usage{
graphic_output(save_path = NULL, file = NULL, resolution = 300)
}
\arguments{
\item{save_path}{If NULL, only draws the graphic in the plot window. Otherwise
specify an output path.}

\item{file}{If NULL, only draws the graphic in the plot window. Otherwise specify
a filename with extension.}

\item{resolution}{DPI resolution.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the dimensions of a graphic produced by \code{\link[=design_graphic]{design_graphic()}}.
}
\examples{
# For default values
gout <- graphic_output()

# Set specific options, the rest will be set to default values
gout <- graphic_output(save_path = "C:/MyFolder/",
                       file      = "MyGraphic.png")

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
