% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_colors.R
\name{graphic_themes}
\alias{graphic_themes}
\alias{add_color_theme}
\alias{get_theme_colors}
\alias{get_theme_base_colors}
\alias{get_theme_font_inside_colors}
\alias{get_theme_font_outside_colors}
\alias{reset_color_themes}
\alias{display_colors}
\alias{display_themes}
\title{Managing Global Color Themes}
\usage{
add_color_theme(
  theme_name,
  base_colors,
  font_inside_colors = c("#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#000000",
    "#000000", "#000000", "#000000", "#000000", "#000000"),
  font_outside_colors = c("#000000", "#000000", "#000000", "#000000", "#000000",
    "#000000", "#000000", "#000000", "#000000", "#000000")
)

get_theme_colors(theme_name)

get_theme_base_colors(theme_name)

get_theme_font_inside_colors(theme_name)

get_theme_font_outside_colors(theme_name)

reset_color_themes(clear_themes = FALSE)

display_colors(theme_name)

display_themes()
}
\arguments{
\item{theme_name}{The name of a globally stored theme.}

\item{base_colors}{The base colors for the segments.}

\item{font_inside_colors}{The individual font colors used when drawing values inside segments
on the corresponding base color.}

\item{font_outside_colors}{The individual font colors used when drawing values outsidr segments
on the corresponding base color.}

\item{clear_themes}{FALSE by default. If TRUE clears the global theme list.}
}
\value{
\code{\link[=add_color_theme]{add_color_theme()}}: Returns modified global theme list.

\code{\link[=get_theme_colors]{get_theme_colors()}}: A list with two vectors of hex color codes.

\code{\link[=get_theme_base_colors]{get_theme_base_colors()}}: A vector of hex color codes.

\code{\link[=get_theme_font_inside_colors]{get_theme_font_inside_colors()}}: A vector of hex color codes.

\code{\link[=get_theme_font_outside_colors]{get_theme_font_outside_colors()}}: A vector of hex color codes.

\code{\link[=reset_color_themes]{reset_color_themes()}}: Default global color list or empty list.

\code{\link[=display_colors]{display_colors()}}: Returns a list of base and font colors.

\code{\link[=display_themes]{display_themes()}}: A list containing all globally stored themes with their color
codew vectors.
}
\description{
\code{\link[=add_color_theme]{add_color_theme()}}: Adds a new color theme containing base and font colors to the
global theme list.

\code{\link[=get_theme_colors]{get_theme_colors()}}: Retrieve a list with two vectors of hex colors from the
globally set up themes containing the base colors for the segments and the
corresponding font colors.

\code{\link[=get_theme_base_colors]{get_theme_base_colors()}}: Retrieve a vector of hex colors from the globally set up themes
containing the base colors for the segments.

\code{\link[=get_theme_font_inside_colors]{get_theme_font_inside_colors()}}: Retrieve a vector of hex colors from the globally
set up themes containing the individual font colors, which are used when values are
drawn inside segments, corresponding to the base segment colors.

\code{\link[=get_theme_font_outside_colors]{get_theme_font_outside_colors()}}: Retrieve a vector of hex colors from the globally
set up themes containing the individual font colors, which are used when values are
drawn outside segments, corresponding to the base segment colors.

\code{\link[=reset_color_themes]{reset_color_themes()}}: Resets the global color themes back to default.

\code{\link[=display_colors]{display_colors()}}: Displays all colors of a theme with the corresponding hex codes.

\code{\link[=display_themes]{display_themes()}}: Display all colors of all themes currently present in the
global environment.
}
\examples{
# Adding and getting color themes
add_color_theme(theme_name  = "tropic",
                base_colors = c("#1B3A2E", "#1F5A3F", "#257F54", "#2FA36A", "#56C07E",
                                "#85D49A", "#B2E4B8", "#D7F0D5", "#ECF8EB", "#F7FCF7"),
                font_inside_colors = c("#FFFFFF", "#FFFFFF", "#FFFFFF", "#FFFFFF", "#000000",
                                        "#000000", "#000000", "#000000", "#000000", "#000000"),
                font_outside_colors = c("#1B3A2E", "#1F5A3F", "#257F54", "#2FA36A", "#56C07E",
                                        "#85D49A", "#B2E4B8", "#D7F0D5", "#ECF8EB", "#F7FCF7"))

get_theme_colors("ocean")

get_theme_base_colors("ocean")

get_theme_font_inside_colors("ocean")

get_theme_font_outside_colors("ocean")

# Clear all globally stored themes in case only self created themes should be stored
reset_color_themes(clear_themes = TRUE)

# Reset global themes to default
reset_color_themes()

# Displaying colors and themes
display_colors("ocean")

display_themes()

}
\seealso{
Graphic functions: \code{\link[=design_graphic]{design_graphic()}}

Global graphic options: \code{\link[=add_color_theme]{add_color_theme()}}, \code{\link[=get_theme_base_colors]{get_theme_base_colors()}}, \code{\link[=get_theme_font_inside_colors]{get_theme_font_inside_colors()}},
\code{\link[=get_theme_font_outside_colors]{get_theme_font_outside_colors()}}, \code{\link[=reset_color_themes]{reset_color_themes()}}

View colors and themes: \code{\link[=display_colors]{display_colors()}}, \code{\link[=display_themes]{display_themes()}}
}
