% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_visuals}
\alias{graphic_visuals}
\title{Graphic Visuals}
\usage{
graphic_visuals(
  font = "Arial",
  color_theme = "ocean",
  color_usage = .qol_options[["graphic_color_usage"]],
  theme_override = list(),
  title_font_color = "#000000",
  footnote_font_color = "#000000",
  primary_axes_font_color = "#000000",
  secondary_axes_font_color = "#000000",
  variable_axes_font_color = "#000000",
  label_font_color = "#000000",
  origin_font_color = "#2B2B2B",
  other_font_color = "#000000",
  title_font_face = "bold",
  footnote_font_face = "plain",
  primary_axes_font_face = "plain",
  secondary_axes_font_face = "plain",
  variable_axes_font_face = "plain",
  value_font_face = "bold",
  label_font_face = "plain",
  origin_font_face = "plain",
  other_font_face = "plain",
  title_alignment = "left",
  footnote_alignment = "left",
  hbar_alignment = "right",
  other_alignment = "left",
  reverse_colors = FALSE,
  primary_axes_color = "#9A9A9A",
  secondary_axes_color = "#9A9A9A",
  variable_axes_color = "#9A9A9A",
  graphic_background_color = "#FFFFFF",
  diagram_background_color = "#FFFFFF",
  graphic_border_color = "#000000",
  diagram_border_color = "#FFFFFF",
  segment_border_color = "#000000",
  line_markers = TRUE,
  line_markers_change = TRUE,
  guiding_lines_y = FALSE,
  guiding_lines_x = FALSE,
  guiding_line_type = "dotted",
  guiding_line_color = "#9A9A9A",
  major_separation_line_type = "solid",
  major_separation_line_color = "#9A9A9A",
  minor_separation_line_type = "dotted",
  minor_separation_line_color = "#9A9A9A",
  segment_line_type = "solid",
  segment_line_color = "#000000",
  remove_small_values = TRUE,
  display_values = TRUE,
  bar_values_inside = TRUE,
  rotate_values = FALSE,
  display_plus_symbol = FALSE,
  label_type = "lines",
  label_group = "auto",
  legend_x_pos = "auto",
  legend_y_pos = "auto",
  legend_columns = 1,
  legend_symbol_size = 0.3,
  origin = "Graphic: qol",
  tooltip_font_color = "auto",
  tooltip_background_color = "auto",
  tooltip_border_color = "auto",
  tooltip_border_width = 2,
  tooltip_background_opacity = 0.95,
  tooltip_x_padding = 8,
  tooltip_y_padding = 4,
  tooltip_corner_radius = 2,
  segment_hover_opacity = 0.85,
  group_hover_color = "#6495ED",
  group_hover_opacity = 0.25
)
}
\arguments{
\item{font}{Set the font to be used for the entire output.}

\item{color_theme}{The color theme to use. Can either be the name of a built-in or
custom theme or a vector of hex codes. Existing color themes can be viewed with
\code{\link[=display_themes]{display_themes()}}.}

\item{color_usage}{A list of numerical vectors that specify which specific colors
from the color scheme are used for which number of segments.}

\item{theme_override}{The override parameter takes in a list consisting of
\code{\link[=override_theme]{override_theme()}} functions.}

\item{title_font_color}{Font color of the title.}

\item{footnote_font_color}{Font color of the footnote.}

\item{primary_axes_font_color}{Font color of the primary axes.}

\item{secondary_axes_font_color}{Font color of the secondary axes.}

\item{variable_axes_font_color}{Font color of the variable axes.}

\item{label_font_color}{Font color of the segment labels.}

\item{origin_font_color}{Font color of the origin text.}

\item{other_font_color}{Font color of every other text element.}

\item{title_font_face}{Font face of the title.}

\item{footnote_font_face}{Font face of the footnote.}

\item{primary_axes_font_face}{Font face of the primary axes.}

\item{secondary_axes_font_face}{Font face of the secondary axes.}

\item{variable_axes_font_face}{Font face of the variable axes.}

\item{value_font_face}{Font face of the values.}

\item{label_font_face}{Font face of the segment labels.}

\item{origin_font_face}{Font face of the origin text.}

\item{other_font_face}{Font face of every other text element.}

\item{title_alignment}{The graphic titles alignment.}

\item{footnote_alignment}{The graphic footnote alignment.}

\item{hbar_alignment}{The alignment of the axes group labels in horizontal bars.}

\item{other_alignment}{Alignment of other elements, like freely positionable textboxes.}

\item{reverse_colors}{FALSE by default. If TRUE reverses the color order.}

\item{primary_axes_color}{The color of the primary axes.}

\item{secondary_axes_color}{The color of the secondary axes.}

\item{variable_axes_color}{The color of the variable axes.}

\item{graphic_background_color}{The background color of entire graphic.}

\item{diagram_background_color}{The background color of the diagram area.}

\item{graphic_border_color}{The border color of the entire graphic.}

\item{diagram_border_color}{The border color of the diagram area.}

\item{segment_border_color}{The border color of each segment.}

\item{line_markers}{TRUE by default. Draws markers in line charts. If FALSE,
doesn't draw markers in line charts.}

\item{line_markers_change}{TRUE by default. Uses different markers per line. If
FALSE, dots are used as markers for all lines.}

\item{guiding_lines_y}{FALSE by default. If TRUE draws guiding lines at the y axes tick
positions.}

\item{guiding_lines_x}{FALSE by default. If TRUE draws guiding lines at the x axes tick
positions.}

\item{guiding_line_type}{Sets the type of guiding lines drawn at the y axes tick positions.
Can be "dashed", "dotted" or "solid".}

\item{guiding_line_color}{The color of the guiding lines drawn at the y axes tick positions.}

\item{segment_line_type}{Sets the type of leading lines from segments to labels. Can be
"dashed", "dotted" or "solid".}

\item{segment_line_color}{Sets the color of the leading lines from segments to labels.}

\item{remove_small_values}{TRUE by default. Doesn't display values in stacked diagrams,
if the corresponding segment is to small and value would go out of bounds. If FALSE,
always displays values.}

\item{display_values}{TRUE by default. Displays the values with the segments. If FALSE,
removes all values and just draws the segments.}

\item{bar_values_inside}{TRUE by default. In grouped bar charts the segment values
will be drawn inside the bars. If FALSE, values will be drawn above/beside the bars.}

\item{rotate_values}{FALSE by default. If TRUE rotates values inside the segments by
90 degrees.}

\item{display_plus_symbol}{FALSE by default. If TRUE displays a + symbol in front of
positive values.}

\item{label_type}{Can be "lines", which connects labels and segments with leading lines,
or "legend", which allows to position a legend separately.}

\item{label_group}{If label_Type is "lines", then this parameter determines above which
group of segments the labels will be drawn.}

\item{legend_x_pos}{Horizontal position of the legend.}

\item{legend_y_pos}{Vertical position of the legend.}

\item{legend_columns}{The number of columns in which the labels should be arranged.
If 0, all labels will be drawn below each other. If max number of segments, then
all labels will be drawn beside each other.}

\item{legend_symbol_size}{The size of the symbol that is drawn to the left of the expression text.}

\item{origin}{A character value that will be written in the bottom right corner of the
graphic.}

\item{tooltip_font_color}{Font color of the tooltips in interactive charts. If set to "auto"
the font color will be black or white depending on the tooltip background color.}

\item{tooltip_background_color}{Background color of the tooltips in interactive charts.
If set to "auto" the background color will match the color of the corresponding segment.}

\item{tooltip_border_color}{Border color of the tooltips in interactive charts.
If set to "auto" the border color will match the color of the corresponding segment.}

\item{tooltip_border_width}{Width of the border of the tooltips in interactive charts.}

\item{tooltip_background_opacity}{Background opacity of the tooltips in interactive charts.
0 is fully transparent, while 1 is fully visible.}

\item{tooltip_x_padding}{The horizontal space of the tooltip texts in interactive charts to
the borders.}

\item{tooltip_y_padding}{The vertical space of the tooltip texts in interactive charts to
the borders.}

\item{tooltip_corner_radius}{Radius of the corners of the tooltips in interactive charts.
0 means sharp corners, tooltip displays as rectangle.}

\item{segment_hover_opacity}{When hovering with the mouse cursor over a segment of an
interactive chart, this parameter determines at how much opacity the segment is displayed.}

\item{group_hover_color}{The background color of segment groups used in interactive charts.}

\item{group_hover_opacity}{When hovering with the mouse cursor over a segment group of an
interactive chart, this parameter determines at how much opacity the segment group background
is displayed.}

\item{separation_line_type}{Sets the type of separation lines drawn between top group
categories. Can be "dashed", "dotted" or "solid".}

\item{separation_line_color}{The color of the separation lines drawn between top group
categories.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the visual appearance of a graphic produced by
\code{\link[=design_graphic]{design_graphic()}}.
}
\examples{
# For default values
gv <- graphic_visuals()

# Set specific options, the rest will be set to default values
gv <- graphic_visuals(title_font_color = "#00FF00",
                      line_markers     = FALSE)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
