% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_options.R
\name{graphic_visuals}
\alias{graphic_visuals}
\title{Graphic Visuals}
\usage{
graphic_visuals(
  font = "Arial",
  title_font_color = "#000000",
  footnote_font_color = "#000000",
  primary_axes_font_color = "#000000",
  secondary_axes_font_color = "#000000",
  variable_axes_font_color = "#000000",
  value_font_color = "#000000",
  label_font_color = "#000000",
  origin_font_color = "#2B2B2B",
  other_font_color = "#000000",
  title_font_face = "bold",
  footnote_font_face = "plain",
  primary_axes_font_face = "plain",
  secondary_axes_font_face = "plain",
  variable_axes_font_face = "plain",
  value_font_face = "bold",
  label_font_face = "plain",
  origin_font_face = "plain",
  other_font_face = "plain",
  title_alignment = "left",
  footnote_alignment = "left",
  hbar_alignment = "right",
  other_alignment = "left",
  color_theme = "ocean",
  reverse_colors = FALSE,
  primary_axes_color = "#2B2B2B",
  secondary_axes_color = "#2B2B2B",
  variable_axes_color = "#2B2B2B",
  graphic_background_color = "#FFFFFF",
  diagram_background_color = "#FFFFFF",
  graphic_border_color = "#000000",
  diagram_border_color = "#FFFFFF",
  segment_border_color = "#000000",
  line_markers = TRUE,
  line_markers_change = TRUE,
  guiding_lines = FALSE,
  guiding_line_type = "dotted",
  guiding_line_color = "#2B2B2B",
  separation_line_type = "dashed",
  separation_line_color = "#2B2B2B",
  segment_line_type = "solid",
  segment_line_color = "#000000",
  remove_small_values = TRUE,
  bar_values_inside = TRUE,
  rotate_values = FALSE,
  label_type = "lines",
  label_group = 3,
  line_offset = 0,
  legend_x_pos = 11,
  legend_y_pos = 5,
  legend_columns = 1,
  top_category_position = "top",
  origin = "Graphic: qol"
)
}
\arguments{
\item{font}{Set the font to be used for the entire output.}

\item{title_font_color}{Font color of the title.}

\item{footnote_font_color}{Font color of the footnote.}

\item{primary_axes_font_color}{Font color of the primary axes.}

\item{secondary_axes_font_color}{Font color of the secondary axes.}

\item{variable_axes_font_color}{Font color of the variable axes.}

\item{value_font_color}{Font color of the values.}

\item{label_font_color}{Font color of the segment labels.}

\item{origin_font_color}{Font color of the origin text.}

\item{other_font_color}{Font color of every other text element.}

\item{title_font_face}{Font face of the title.}

\item{footnote_font_face}{Font face of the footnote.}

\item{primary_axes_font_face}{Font face of the primary axes.}

\item{secondary_axes_font_face}{Font face of the secondary axes.}

\item{variable_axes_font_face}{Font face of the variable axes.}

\item{value_font_face}{Font face of the values.}

\item{label_font_face}{Font face of the segment labels.}

\item{origin_font_face}{Font face of the origin text.}

\item{other_font_face}{Font face of every other text element.}

\item{title_alignment}{The graphic titles alignment.}

\item{footnote_alignment}{The graphic footnote alignment.}

\item{hbar_alignment}{The alignment of the axes group labels in horizontal bars.}

\item{other_alignment}{Alignment of other elements, like freely positionable textboxes.}

\item{color_theme}{The color theme used for the segments.}

\item{reverse_colors}{FALSE by default. If TRUE reverses the color order.}

\item{primary_axes_color}{The color of the primary axes.}

\item{secondary_axes_color}{The color of the secondary axes.}

\item{variable_axes_color}{The color of the variable axes.}

\item{graphic_background_color}{The background color of entire graphic.}

\item{diagram_background_color}{The background color of the diagram area.}

\item{graphic_border_color}{The border color of the entire graphic.}

\item{diagram_border_color}{The border color of the diagram area.}

\item{segment_border_color}{The border color of each segment.}

\item{line_markers}{TRUE by default. Draws markers in line charts. If FALSE,
doesn't draw markers in line charts.}

\item{line_markers_change}{TRUE by default. Uses different markers per line. If
FALSE, dots are used as markers for all lines.}

\item{guiding_lines}{FALSE by default. If TRUE draws guiding lines at the y axes tick
positions.}

\item{guiding_line_type}{Sets the type of guiding lines drawn at the y axes tick positions.
Can be "dashed", "dotted" or "solid".}

\item{guiding_line_color}{The color of the guiding lines drawn at the y axes tick positions.}

\item{separation_line_type}{Sets the type of separation lines drawn between top group
categories. Can be "dashed", "dotted" or "solid".}

\item{separation_line_color}{The color of the separation lines drawn between top group
categories.}

\item{segment_line_type}{Sets the type of leading lines from segments to labels. Can be
"dashed", "dotted" or "solid".}

\item{segment_line_color}{Sets the color of the leading lines from segments to labels.}

\item{remove_small_values}{TRUE by default. Doesn't display values if the corresponding
segment is to small and value would go out of bounds. If FALSE, always displays values.}

\item{bar_values_inside}{TRUE by default. In grouped bar charts the segment values
will be drawn inside the bars. If FALSE, values will be drawn above/beside the bars.}

\item{rotate_values}{FALSE by default. If TRUE rotates values inside the segments by
90 degrees.}

\item{label_type}{Can be "lines", which connects labels and segments with leading lines,
or "legend", which allows to position a legend separately.}

\item{label_group}{If label_Type is "lines", then this parameter determines above which
group of segments the labels will be drawn.}

\item{line_offset}{Offset in the height of the leading lines in cm. If 0, all leading lines
will end at the same height, meaning the labels will be drawn at the same height. If
an offset is set, all lines after the first one will be shorter by this amount * number
of the line. Enables the labels to be drawn in steps, if e.g. the labels have a long
text which would overlap, if they where on the same height.}

\item{legend_x_pos}{Horizontal position of the legend.}

\item{legend_y_pos}{Vertical position of the legend.}

\item{legend_columns}{The number of columns in which the labels should be arranged.
If 0, all labels will be drawn below each other. If max number of segments, then
all labels will be drawn beside each other.}

\item{top_category_position}{If "top", the top group categories will be drawn above the
diagram, otherwise if "bottom" they will be drawn belowthe diagram.}

\item{origin}{A character value that will be written in the bottom right corner of the
graphic.}
}
\value{
Returns a list of named graphic options.
}
\description{
Set different options regarding the visual appearance of a graphic produced by
\code{\link[=design_graphic]{design_graphic()}}.
}
\examples{
# For default values
gv <- graphic_visuals()

# Set specific options, the rest will be set to default values
gv <- graphic_visuals(title_font_color = "#00FF00",
                      line_markers     = FALSE)

}
\seealso{
The main graphic function: \code{\link[=design_graphic]{design_graphic()}}

Other graphic options:

Additional graphic functions: \code{\link[=add_textbox]{add_textbox()}}
}
