% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_random_missingness.R
\name{inject_na}
\alias{inject_na}
\title{Inject Random Missingness}
\usage{
inject_na(data = ., prop = 0.1)
}
\arguments{
\item{data}{}

\item{prop}{}
}
\description{
Inject Random Missingness
}
