% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_diagrams.R
\name{main_diagram}
\alias{main_diagram}
\alias{dg_vbars}
\title{Generate The Main Diagram Area}
\usage{
dg_vbars(...)
}
\arguments{
\item{...}{Arguments passed on by \code{\link[=design_graphic]{design_graphic()}}.}
}
\value{
Returns a grid::gTree object.
}
\description{
The listed functions are the built in diagram types which ship with this package.
These functions don't work on their own and have to be plugged into the "diagram"
parameter of \code{\link[=design_graphic]{design_graphic()}}.

\code{\link[=dg_vbars]{dg_vbars()}}: Grouped vertical bars.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 25" = 18:24,
    "25 to under 55" = 25:54,
    "55 to under 65" = 55:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Total"  = 1:2,
    "Male"   = 1,
    "Female" = 2)

# Design grouped vertical bar chart
qol_graphic <- my_data |>
     design_graphic(axes_variables = "sex",
                    segments       = "age",
                    values         = weight,
                    diagram        = dg_vbars,
                    formats        = list(sex = sex.,
                                          age = age.))

}
