% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_noisy_data.R
\name{naugment}
\alias{naugment}
\title{Augment Noisy Data}
\usage{
naugment(
  data,
  tree_n = 150,
  initial_n_prop = 0.1,
  digits = 0,
  aug_n = nrow(data),
  replace = FALSE
)
}
\arguments{
\item{data}{}

\item{tree_n}{}

\item{initial_n_prop}{}

\item{digits}{}

\item{aug_n}{}

\item{replace}{}
}
\description{
Augment Noisy Data
}
