% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{output_graphic}
\alias{output_graphic}
\title{Output Graphic To Device}
\usage{
output_graphic(graphic_object, dimensions, output)
}
\arguments{
\item{graphic_object}{The complete graphic to be drawn or exported.}

\item{dimensions}{qol package dimensions options.}

\item{output}{qol package output options.}
}
\value{
Returns the input graphic object.
}
\description{
Draw or output a graphic to a desired destination.
}
