% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qol_options}
\alias{qol_options}
\alias{set_print}
\alias{get_print}
\alias{set_monitor}
\alias{get_monitor}
\alias{set_na.rm}
\alias{get_na.rm}
\alias{set_output}
\alias{get_output}
\title{Set Global Print Option}
\usage{
set_print(...)

get_print()

set_monitor(...)

get_monitor()

set_na.rm(...)

get_na.rm()

set_output(...)

get_output()
}
\arguments{
\item{...}{Put in TRUE or FALSE to activate or deactivate the option.}
}
\value{
\code{\link[=set_print]{set_print()}}: Change global print option.

\code{\link[=get_print]{get_print()}}: TRUE or FALSE.

\code{\link[=set_monitor]{set_monitor()}}: Change global monitor option.

\code{\link[=get_monitor]{get_monitor()}}: TRUE or FALSE.

\code{\link[=set_na.rm]{set_na.rm()}}: Change global na.rm option.

\code{\link[=get_na.rm]{get_na.rm()}}: TRUE or FALSE.

\code{\link[=set_output]{set_output()}}: Change global output option.

\code{\link[=get_output]{get_output()}}: Current output option as character.
}
\description{
\code{\link[=set_print]{set_print()}}: Set the print option globally for the tabulation and export to
Excel functions.

\code{\link[=get_print]{get_print()}}: Get the globally stored print option.

\code{\link[=set_monitor]{set_monitor()}}: Set the monitor option globally for the heavier functions which are
able to show how they work internally.

\code{\link[=get_monitor]{get_monitor()}}: Get the globally stored monitor option.

\code{\link[=set_na.rm]{set_na.rm()}}: Set the na.rm option globally for each function which can remove
NA values.

\code{\link[=get_na.rm]{get_na.rm()}}: Get the globally stored na.rm option.

\code{\link[=set_output]{set_output()}}: Set the output option globally for each function that can output
results to "console", "text", "excel" or "excel_nostyle".

\code{\link[=get_output]{get_output()}}: Get the globally stored output option.
}
\examples{
set_print(FALSE)
set_print(TRUE)

get_print()

set_monitor(TRUE)
set_monitor(FALSE)

get_monitor()

set_na.rm(TRUE)
set_na.rm(FALSE)

get_na.rm()

set_output("excel")

get_output()

}
