% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{qol_options}
\alias{qol_options}
\alias{set_style_options}
\alias{reset_style_options}
\alias{get_style_options}
\title{Set Global Styling Options For Excel Workbooks}
\usage{
set_style_options(...)

reset_style_options()

get_style_options()
}
\arguments{
\item{...}{Put in any styling option from \code{\link[=excel_output_style]{excel_output_style()}} or \code{\link[=number_format_style]{number_format_style()}}
with the new value.}
}
\value{
\code{\link[=set_style_options]{set_style_options()}}: Returns modified global styling options.

\code{\link[=reset_style_options]{reset_style_options()}}: Returns default global styling options.

\code{\link[=get_style_options]{get_style_options()}}: Printed output.
}
\description{
Modify Styling options for Excel workbooks. Available parameters can be seen in
\code{\link[=excel_output_style]{excel_output_style()}} or \code{\link[=number_format_style]{number_format_style()}}.

\code{\link[=set_style_options]{set_style_options()}} sets the styling options for Excel workbooks globally.These
options are used by all tabulation and output functions, which are capable of
exporting styled outputs.

\code{\link[=reset_style_options]{reset_style_options()}} resets global style options to the default parameters.

\code{\link[=get_style_options]{get_style_options()}} prints out the currently set global styling options.
}
\examples{
set_style_options(save_path    = "C:/My Projects/",
                  sum_decimals = 8)

reset_style_options()

get_style_options()

}
