% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_helpers.R
\name{round_values}
\alias{round_values}
\title{Round Values With Half Rounded Up}
\usage{
round_values(values, digits = 0)
}
\arguments{
\item{values}{Numeric values to round.}

\item{digits}{The number of decimal places the values should be rounded to.}
}
\value{
Returns rounded values
}
\description{
This function rounds values according to DIN 1333 (round half up).
}
\examples{
round_numbers1 <- round_values(c(-0.5, -0.4, 0.1, 0.49, 0.5, 1.5, 2.5, 3.2))
round_numbers2 <- round_values(c(-0.5, -0.49, 0.17, 0.499, 0.51, 1.549, 2.51, 3.25),
                               digits = 1)

}
