% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_master.R
\name{run_scripts}
\alias{run_scripts}
\alias{run_script}
\alias{run_folder}
\alias{run_project}
\alias{run_project_parallel}
\title{Run 'R' Scripts Via File Dialog}
\usage{
run_script(path = NULL)

run_folder(path = NULL)

run_project(path = NULL)

run_project_parallel(path = NULL)
}
\arguments{
\item{path}{The folder in which the file selection dialog should start. If
left out, the folder of the currently opened script in RStudio is used.}
}
\value{
Returns TRUE or FALSE.
}
\description{
Opens an RStudio file selection dialog to pick a single script/folder or a
whole folder structure and runs it after confirmation. The dialog starts in the
folder of the currently opened script in RStudio.

Single files and entire folders can be silenced (code will not be executed) by
adding a "_" to the end of the file or folder name.

\code{\link[=run_script]{run_script()}}: Runs the selected script.

\code{\link[=run_folder]{run_folder()}}: Runs all scripts in the selected folder.

\code{\link[=run_project]{run_project()}}: Runs all scripts found in a folder structure, including all
scripts within subfolders, in sequential order.

\code{\link[=run_project_parallel]{run_project_parallel()}}: Runs all scripts found in a folder structure, including
all scripts within subfolders. Folders, which are marked with an "!" at the end
of their folder name, are run as background jobs in parallel, all other folders
are run sequentially.
}
\details{
Example structure for a larger project

Project folder
\itemize{
\item Subfolder1
\itemize{
\item Script1.R
\item Script2.R
\item ...
}
\item Subfolder2
\itemize{
\item ...
}
\item ...
\item Master.Rmd
}

If you open "Master.Rmd" in such a structure (the file can be completely empty)
and simply run one of the commands in the console, a selection window opens
which directly shows the path where the "Master.Rmd" file is located.
Depending on the command, you can then select and run either a single script,
all scripts in a folder, or all scripts in all folders of the project at once.

Note: If the master file has the extension "*.R", it is also picked up when
running the project folder with \code{\link[=run_project]{run_project()}}. If it also contains
code which runs files in the subfolders, things will be executed twice.
Therefore it is better to use a different file type or to not add any
code in it.
}
\seealso{
Build master file: \code{\link[=build_master]{build_master()}}
}
