% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{segment_labels}
\alias{segment_labels}
\alias{direct_vertical_labels}
\alias{setup_legend}
\title{Generate Segment Labels And Legends}
\usage{
direct_vertical_labels(diagram_info, arguments)

setup_legend(diagram_info, visuals, dimensions)
}
\arguments{
\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{arguments}{Argument list passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{visuals}{Visual parameters set with \code{\link[=graphic_visuals]{graphic_visuals()}}.}

\item{dimensions}{Dimension parameters set with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.}
}
\value{
\code{\link[=direct_vertical_labels]{direct_vertical_labels()}}: Returns a grid::gList.

\code{\link[=setup_legend]{setup_legend()}}: Returns a grid::gList and adjusted diagram dimensions.
}
\description{
\code{\link[=direct_vertical_labels]{direct_vertical_labels()}}: Set up segment labels which are connected directly
to the segments by a line.

\code{\link[=setup_legend]{setup_legend()}}: Set up a legend for the segment labels. The legend consists
of a colored shape followed by th label of the individual segment. The legend
is drawn in single or multi column layout.
}
