% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{segment_labels}
\alias{segment_labels}
\alias{direct_vertical_labels}
\title{Generate Segment Labels And Legends}
\usage{
direct_vertical_labels(diagram_info, arguments)
}
\arguments{
\item{diagram_info}{The list of measurements generated by
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}.}

\item{arguments}{Argument list passed passed on by \code{\link[=design_graphic]{design_graphic()}}.}
}
\value{
\code{\link[=direct_vertical_labels]{direct_vertical_labels()}}: Returns a grid::gList.
}
\description{
\code{\link[=direct_vertical_labels]{direct_vertical_labels()}}: Set up segment labels which are connected directly
to the segments by a line.
}
