% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{textboxes}
\alias{textboxes}
\alias{add_textbox}
\alias{get_text_width}
\alias{get_text_height}
\title{Add Textboxes As Graphical Object}
\usage{
add_textbox(
  text,
  x_pos = .qol_options[["graphic_dimensions"]][["margins"]],
  y_pos = .qol_options[["graphic_dimensions"]][["margins"]],
  width = .qol_options[["graphic_dimensions"]][["textbox_width"]],
  alignment = .qol_options[["graphic_visuals"]][["other_alignment"]],
  font = .qol_options[["graphic_visuals"]][["font"]],
  font_color = .qol_options[["graphic_visuals"]][["other_font_color"]],
  font_size = .qol_options[["graphic_dimensions"]][["other_font_size"]],
  font_face = .qol_options[["graphic_visuals"]][["other_font_face"]],
  line_height = .qol_options[["graphic_dimensions"]][["line_height"]],
  rotation = 0,
  name = "textbox",
  draw = FALSE
)

get_text_width(text, type, dimensions, visuals, unit = "native")

get_text_height(text, type, dimensions, visuals, unit = "native")
}
\arguments{
\item{text}{The text that should be displayed.}

\item{x_pos}{X starting position of the textbox in cm.}

\item{y_pos}{Y starting position of the textbox in cm. 0 position is at the top.}

\item{width}{Width of the textbox in cm.}

\item{alignment}{Horizontal text alignment.}

\item{font}{Name of the font to be used.}

\item{font_color}{Font color as hex code.}

\item{font_size}{Font size.}

\item{font_face}{Valid values are "plain", "bold", "italic", "oblique", and "bold.italic".}

\item{line_height}{The height of a single text line.}

\item{rotation}{Rotates the text by a number of degrees.}

\item{name}{The internal name of the textbox with which it can be identified.}

\item{draw}{FALSE by default. If TRUE, directly draws the textbox onto the canvas.}

\item{type}{The type of text to be measured. Can be: title, footnote, primary_axes,
secondary_axes, variable_axes, value, label, origin or other.}

\item{dimensions}{Dimension parameters set with \code{\link[=graphic_dimensions]{graphic_dimensions()}}.}

\item{visuals}{Visual parameters set with \code{\link[=graphic_visuals]{graphic_visuals()}}.}

\item{unit}{The unit to measure widths and heights in. Used are "native" for measuring
inside the inner canvas. Otherwise use "cm".}
}
\value{
Returns a grid::textGrob object.

\code{\link[=get_text_width]{get_text_width()}}: Returns a numeric width.

\code{\link[=get_text_height]{get_text_height()}}: Returns a numeric height.
}
\description{
\code{\link[=add_textbox]{add_textbox()}}: Create a textbox as graphical object. The dimensions and overall visual
appearance is flexible. Text wraps automatically, if the provided text is wider than the
textbox itself. Textboxes can be plugged into the "add_texts" parameter of \code{\link[=design_graphic]{design_graphic()}}.

\code{\link[=get_text_width]{get_text_width()}}: Get the width of a text while applying individual styling.

\code{\link[=get_text_height]{get_text_height()}}: Get the height of a text while applying individual styling.
}
\examples{
# Example data frame
my_data <- dummy_data(100)

# Formats
age. <- discrete_format(
    "Total"          = 0:100,
    "under 18"       = 0:17,
    "18 to under 65" = 18:64,
    "65 and older"   = 65:100)

sex. <- discrete_format(
    "Male"   = 1,
    "Female" = 2)

# Add individual texts to a graphic
my_data |>
     design_graphic(axes_variables = "age",
                    segments       = "sex",
                    values         = weight,
                    diagram        = dg_vbars,
                    add_texts = list(add_textbox("Hello",  3.5, 3),
                                     add_textbox("World!", 6,   5)))

}
