% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{textboxes}
\alias{textboxes}
\alias{add_textbox}
\alias{wrap_single_text}
\alias{wrap_text_vector}
\alias{add_title}
\alias{add_footnote}
\alias{add_graphic_origin}
\alias{register_windows_font}
\title{Add Textboxes As Graphical Object}
\usage{
add_textbox(
  text,
  x_pos = .qol_options[["graphic_dimensions"]][["margins"]],
  y_pos = .qol_options[["graphic_dimensions"]][["margins"]],
  width = .qol_options[["graphic_dimensions"]][["graphic_width"]],
  alignment = .qol_options[["graphic_visuals"]][["other_alignment"]],
  font = .qol_options[["graphic_visuals"]][["font"]],
  font_color = .qol_options[["graphic_visuals"]][["other_font_color"]],
  font_size = .qol_options[["graphic_dimensions"]][["other_font_size"]],
  font_face = .qol_options[["graphic_visuals"]][["other_font_face"]],
  name = "textbox",
  draw = FALSE
)

wrap_single_text(text_line, textbox_width, font, font_size, font_face)

wrap_text_vector(text_lines, textbox_width, font, font_size, font_face)

add_title(
  text,
  dimensions = .qol_options[["graphic_dimensions"]],
  visuals = .qol_options[["graphic_visuals"]],
  draw = FALSE
)

add_footnote(
  text,
  dimensions = .qol_options[["graphic_dimensions"]],
  visuals = .qol_options[["graphic_visuals"]],
  draw = FALSE
)

add_graphic_origin(
  dimensions = .qol_options[["graphic_dimensions"]],
  visuals = .qol_options[["graphic_visuals"]],
  draw = FALSE
)

register_windows_font(font)
}
\arguments{
\item{text}{The text that should be displayed.}

\item{x_pos}{X starting position of the textbox in cm.}

\item{y_pos}{Y starting position of the textbox in cm. 0 position is at the top.}

\item{width}{Width of the textbox in cm.}

\item{alignment}{Horizontal text alignment.}

\item{font}{Name of the font to be used.}

\item{font_color}{Font color as hex code.}

\item{font_size}{Font size.}

\item{font_face}{Valid values are "plain", "bold", "italic", "oblique", and "bold.italic".}

\item{name}{The internal name of the textbox with which it can be identified.}

\item{draw}{FALSE by default. If TRUE, directly draws the textbox onto the canvas.}

\item{text_line}{A single line of text to wrap.}

\item{textbox_width}{Width of the textbox in cm.}

\item{text_lines}{A vector of texts to wrap.}

\item{dimensions}{qol package dimensions options.}

\item{visuals}{qol package visuals options.}
}
\value{
Returns a grid::textGrob object.

Returns wrapped character.

Returns wrapped character vector.
}
\description{
\code{\link[=add_textbox]{add_textbox()}}: Create a textbox as graphical object. The dimensions and overall visual appearance
is flexible. Text wraps automatically, if the provided text is wider than the
textbox itself.

\code{\link[=wrap_single_text]{wrap_single_text()}}: Wrap text based on a provided width.

\code{\link[=wrap_text_vector]{wrap_text_vector()}}: Wrap every single element of a character vector individually.

\code{\link[=add_title]{add_title()}}: Create a textbox as graphical object. A wrapper to easily create
a textbox at the top of the graphic as title.

\code{\link[=add_footnote]{add_footnote()}}: Create a textbox as graphical object. A wrapper to easily create
a textbox at the bottom of the graphic as footnote.

\code{\link[=add_graphic_origin]{add_graphic_origin()}}: Create a textbox as graphical object. A wrapper to easily create
a textbox at the bottom right of the graphic as information who created the graphic.

\code{\link[=register_windows_font]{register_windows_font()}}: Windows only. Registers the font to be used.
}
\seealso{
High level graphic functions: \code{\link[=design_graphic]{design_graphic()}}

Mid level graphic functions:

Add textboxes: \code{\link[=add_textbox]{add_textbox()}}, \code{\link[=add_title]{add_title()}}, \code{\link[=add_footnote]{add_footnote()}}, \code{\link[=add_graphic_origin]{add_graphic_origin()}}

Viewport: \code{\link[=setup_main_canvas]{setup_main_canvas()}}, \code{\link[=is_viewport_pushed]{is_viewport_pushed()}}

Dimensions: \code{\link[=get_available_width]{get_available_width()}}, \code{\link[=get_available_height]{get_available_height()}}
}
