% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_helpers.R
\name{viewport}
\alias{viewport}
\alias{setup_nested_diagram_viewport}
\alias{back_to_the_root}
\title{Set Up Viewports To Draw On}
\usage{
setup_nested_diagram_viewport(arguments, diagram_info)

back_to_the_root(to_main = TRUE)
}
\arguments{
\item{arguments}{Argument list passed on by \code{\link[=design_graphic]{design_graphic()}}.}

\item{diagram_info}{The list of measurements generated by \code{\link[=get_diagram_dimensions]{get_diagram_dimensions()}}.}

\item{to_main}{TRUE by default. Pops out of all viewports but the main canvas.}
}
\value{
Returns the grid::current.vpPath.
}
\description{
\code{\link[=setup_nested_diagram_viewport]{setup_nested_diagram_viewport()}}: Setup the main diagram viewport based on
the list of arguments passed on by \code{\link[=design_graphic]{design_graphic()}}. This viewport contains
the main diagram area as well as the axes. Additionally sets up the main diagram
area viewport which only contains the main drawing area.

\code{\link[=back_to_the_root]{back_to_the_root()}}: Pops out of all viewports.
}
