% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_anova.R
\name{wrap_anova}
\alias{wrap_anova}
\alias{anova}
\title{Analysis of Variance (ANOVA) Wrapper}
\usage{
wrap_anova(
  data,
  outcome = NULL,
  between_vars = NULL,
  within_vars = NULL,
  rm = "within",
  value = "value",
  id = NULL,
  covariates = NULL,
  interactions = FALSE,
  rm_correction = "auto",
  posthoc_correction = "tukey",
  conf_level = 0.95,
  es_type = "hedges",
  emm_weights = TRUE,
  filename = NULL,
  study_name = NULL,
  assumptions = FALSE,
  digits = 3,
  summary_type = 1,
  report_es = c("etaSqP")
)

anova(
  data,
  outcome = NULL,
  between_vars = NULL,
  within_vars = NULL,
  rm = "within",
  value = "value",
  id = NULL,
  covariates = NULL,
  interactions = FALSE,
  rm_correction = "auto",
  posthoc_correction = "tukey",
  conf_level = 0.95,
  es_type = "hedges",
  emm_weights = TRUE,
  filename = NULL,
  study_name = NULL,
  assumptions = FALSE,
  digits = 3,
  summary_type = 1,
  report_es = c("etaSqP")
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcome}{A required character indicating the outcome/dependent variable
when running one-way ANOVA.}

\item{between_vars}{Required character(s) indicating the factor grouping
variable(s) when running one-way ANOVA. Optional when running
repeated-measures ANOVA.}

\item{within_vars}{Required characters indicating the within-subjects or
repeated variables when running repeated-measures ANOVA.}

\item{rm}{An optional character to specify the name of the repeated measures.}

\item{value}{An optional character to specify the name of the scores of the
repeated measures.}

\item{id}{A required character indicating the observation ID variable when
running repeated-measures ANOVA.}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{es_type}{Character indicating which effect size to return for pairwise
comparisons. Options include \code{"cohen"} for Cohen's d, \code{"hedges"}
for Hedges' g, or \code{"glass"} for Glass's delta. The default is Hedges'
g.}

\item{emm_weights}{A logical value indicating whether to weigh each cell
equally or according to the cell frequency.}

\item{filename}{Location to save output file as excel workbook if specified.}

\item{study_name}{Relevant only if \code{filename} is specified. An optional
character string indicating the name of the study.}

\item{assumptions}{Relevant only if \code{filename} is specified. A logical
indicating whether to add an ANOVA assumptions sheet. The default is
\code{FALSE}.}

\item{digits}{Relevant only if \code{filename} is specified. A numeric value
indicating the number of digits to round to.}

\item{summary_type}{Relevant only if \code{filename} is specified. A numeric
value indicating the format of the ANOVA summary page in the workbook.
\code{summary_type = 1} combines the between and within-subjects ANOVA
summaries together. \code{summary_type = 2} keeps the between and
within-subjects ANOVA summaries separate.}

\item{report_es}{Relevant only if \code{filename} is specified. A vector of
strings indicating which effect sizes to report. Options include
\code{"ges"} for generalized eta-squared, \code{"etaSq"} for eta-squared,
\code{"etaSqP"} for partial eta-squared, \code{"omegaSq"} for
omega-squared, and \code{"cohen_f"} for Cohen's f. Specify \code{NULL} to
not report any effect size.}

\item{interactionss}{Either a logical or vector of characters representing
the interactions effects of interest. If specified as \code{TRUE}, all
possible interactionss between the groups are run (not recommended). If
specified as \code{FALSE}, no interactions terms are run. If specified with
interactions
terms, e.g., \code{c("treatment\*gender", "treatment\*age_group",
"treatment\*gender\*age_group")}, the only the specified interactions terms
are run. The default is \code{FALSE}.}

\item{ph_method}{The post-hoc pairwise comparison correction method. Options
include \code{"tukey"}, \code{"scheffe"} , \code{"sidak"},
\code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and \code{"none"}.
The default is \code{"tukey"}. See below for additional information.}
}
\description{
A wrapper function to automate running and extracting the
  results of Analysis of Variance (ANOVA).
}
\details{
The current function is limited to running between and within
  subjects ANOVA with factorial or mixed effects models. Pairwise comparison
  effect sizes are limited to Cohen's d, Hedges' g, and Glass's delta.
}
