% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_wrapper.R
\name{wrap_anova}
\alias{wrap_anova}
\title{Analysis of Variance (ANOVA) Wrapper}
\usage{
wrap_anova(
  data,
  outcome,
  groups,
  covariates = NULL,
  interaction = FALSE,
  ph_method = "tukey",
  conf_level = 0.95,
  es_type = "hedges",
  g_correction = "standard"
)
}
\arguments{
\item{data}{A dataframe object. This should be a structured dataset where
each column represents a variable and each row represents an observation.}

\item{outcome}{Character indicating the outcome/dependent variable.}

\item{groups}{Character indicating the factor grouping variable.}

\item{covariates}{An optional vector of variable names representing
covariates of interest.}

\item{interaction}{Either a logical or vector of characters representing the
interaction effects of interest. If specified as \code{TRUE}, all possible
interactions between the groups are run (not recommended). If specified as
\code{FALSE}, no interaction terms are run. If specified with interaction
terms, e.g., \code{c("treatment\*gender", "treatment\*age_group",
"treatment\*gender\*age_group")}, the only the specified interaction terms
are run. The default is \code{FALSE}.}

\item{ph_method}{The post-hoc pairwise comparison correction method. Options
include \code{"tukey"}, \code{"scheffe"} , \code{"sidak"},
\code{"bonferroni"}, \code{"dunnettx"}, \code{"mvt"}, and \code{"none"}.
The default is \code{"tukey"}. See below for additional information.}

\item{conf_level}{Numeric value reflecting the confidence interval level. The
default is 0.95 for 95% confidence interval.}

\item{es_type}{Character indicating which effect size to return. Options
include \code{"cohen"} for Cohen's d, \code{"hedges"} for Hedges' g, or
\code{"glass"} for Glass's delta. The default is Hedges' g.}

\item{g_correction}{Character indicating which Hedges' g correction to apply.
Options include \code{"standard"} for regular Hedges' g correction and
\code{"log"} to apply logarithmic Hedges' g correction. The default is
standard.}
}
\description{
A wrapper for running Analysis of Variance (ANOVA) with or
without covariates.
}
