% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_anova_workbook.R
\name{write_anova}
\alias{write_anova}
\title{Write ANOVA Report Workbook}
\usage{
write_anova(
  anova_object,
  filename,
  study_name = NULL,
  assumptions = FALSE,
  digits = 3,
  summary_type = 1,
  report_es = NULL
)
}
\arguments{
\item{anova_object}{The ANOVA object from \code{wrap_anova}.}

\item{filename}{The filepath to save the workbook.}

\item{study_name}{An optional character string indicating the name of the
study.}

\item{assumptions}{A logical indicating whether to add an ANOVA assumptions
sheet. The default is \code{FALSE}.}

\item{digits}{A numeric value indicating the number of digits to round to.}

\item{summary_type}{A numeric value indicating the format of the ANOVA
summary page in the workbook. \code{summary_type = 1} combines the between
and within-subjects ANOVA summaries together. \code{summary_type = 2} keeps
the between and within-subjects ANOVA summaries separate.}

\item{report_es}{A vector of strings indicating which effect sizes to report.
Options include \code{"ges"} for generalized eta-squared, \code{"etaSq"}
for eta-squared, \code{"etaSqP"} for partial eta-squared, \code{"omegaSq"}
for omega-squared, and \code{"cohen_f"} for Cohen's f. Specify \code{NULL}
to not report any effect size.}
}
\description{
Write ANOVA Report Workbook
}
